Split client and jobs files into smaller modules
