+++ orchestrator/src/client/api/client.ts
@@ -1,2198 +1,8 @@
-/**
- * API client for the orchestrator backend.
- */
-
-import { redirectToSignIn } from "@client/lib/auth-navigation";
-import type { UpdateSettingsInput } from "@shared/settings-schema";
-import type {
-  ApiResponse,
-  ApplicationStage,
-  ApplicationTask,
-  AppSettings,
-  BackupInfo,
-  BranchInfo,
-  CreateJobNoteInput,
-  DemoInfoResponse,
-  DesignResumeDocument,
-  DesignResumeExportResponse,
-  DesignResumeJson,
-  DesignResumePatchRequest,
-  DesignResumePdfResponse,
-  DesignResumeStatusResponse,
-  Job,
-  JobActionRequest,
-  JobActionResponse,
-  JobActionStreamEvent,
-  JobChatMessage,
-  JobChatStreamEvent,
-  JobChatThread,
-  JobListItem,
-  JobNote,
-  JobOutcome,
-  JobSource,
-  JobsListResponse,
-  JobsRevisionResponse,
-  JobTracerLinksResponse,
-  LocationMatchStrictness,
-  LocationSearchScope,
-  ManualJobDraft,
-  ManualJobFetchResponse,
-  ManualJobInferenceResponse,
-  PipelineProgressState,
-  PipelineRun,
-  PipelineRunInsights,
-  PipelineStatusResponse,
-  PostApplicationAction,
-  PostApplicationActionResponse,
-  PostApplicationInboxItem,
-  PostApplicationProvider,
-  PostApplicationProviderActionResponse,
-  PostApplicationRouterStageTarget,
-  PostApplicationSyncRun,
-  ProfileStatusResponse,
-  ResumeProfile,
-  ResumeProjectCatalogItem,
-  SearchTermsSuggestionResponse,
-  StageEvent,
-  StageEventMetadata,
-  StageTransitionTarget,
-  TracerAnalyticsResponse,
-  TracerReadinessResponse,
-  UpdateJobNoteInput,
-  ValidationResult,
-  VisaSponsor,
-  VisaSponsorSearchResponse,
-  VisaSponsorStatusResponse,
-} from "@shared/types";
-import { formatUserFacingError } from "@/client/lib/error-format";
-import { queryClient } from "@/client/lib/queryClient";
-import {
-  bucketQueryLength,
-  getAnalyticsRequestHeaders,
-  trackProductEvent,
-} from "@/lib/analytics";
-import { showDemoBlockedToast, showDemoSimulatedToast } from "@/lib/demo-toast";
-
-const API_BASE = "/api";
-
-class ApiClientError extends Error {
-  requestId?: string;
-  status?: number;
-  code?: string;
-
-  constructor(
-    message: string,
-    options?: { requestId?: string; status?: number; code?: string },
-  ) {
-    super(message);
-    this.name = "ApiClientError";
-    this.requestId = options?.requestId;
-    this.status = options?.status;
-    this.code = options?.code;
-  }
-}
-
-type LegacyApiResponse<T> =
-  | {
-      success: true;
-      data?: T;
-      message?: string;
-    }
-  | {
-      success: false;
-      error?: string;
-      message?: string;
-      details?: unknown;
-    };
-
-type StreamSseInput =
-  | JobActionRequest
-  | { content: string; selectedNoteIds?: string[]; stream: true }
-  | { selectedNoteIds?: string[]; stream: true };
-
-export type CodexAuthStatusResponse = {
-  authenticated: boolean;
-  username: string | null;
-  validationMessage: string | null;
-  flowStatus: string;
-  loginInProgress: boolean;
-  verificationUrl: string | null;
-  userCode: string | null;
-  startedAt: string | null;
-  expiresAt: string | null;
-  flowMessage: string | null;
-};
-
-export type AuthCredentials = {
-  username: string;
-  password: string;
-};
-
-export type AuthUser = {
-  id: string;
-  username: string;
-  displayName: string | null;
-  isSystemAdmin: boolean;
-  isDisabled: boolean;
-  workspaceId: string;
-  workspaceName: string;
-  createdAt: string;
-  updatedAt: string;
-};
-
-export type AuthBootstrapStatus = {
-  setupRequired: boolean;
-};
-
-type StoredLegacyAuthCredentials = AuthCredentials & {
-  storedAt?: number;
-};
-
-const LEGACY_SESSION_AUTH_KEY = "jobops.basicAuthCredentials";
-const LEGACY_SESSION_JWT_KEY = "jobops.jwtToken";
-const SESSION_AUTH_TOKEN_KEY = "jobops.authToken";
-const LEGACY_SESSION_AUTH_TTL_MS = 5 * 60 * 1000;
-
-function decodeBase64UrlJsonSegment(
-  segment: string,
-): Record<string, unknown> | null {
-  try {
-    const normalized = segment.replace(/-/g, "+").replace(/_/g, "/");
-    const padded = normalized.padEnd(
-      normalized.length + ((4 - (normalized.length % 4)) % 4),
-      "=",
-    );
-    const decoded = globalThis.atob(padded);
-    const parsed = JSON.parse(decoded) as unknown;
-    return parsed && typeof parsed === "object" && !Array.isArray(parsed)
-      ? (parsed as Record<string, unknown>)
-      : null;
-  } catch {
-    return null;
-  }
-}
-
-function getTenantIdFromAuthToken(token: string | null): string | null {
-  const payloadSegment = token?.split(".")[1];
-  if (!payloadSegment) return null;
-  const payload = decodeBase64UrlJsonSegment(payloadSegment);
-  const tenantId = payload?.tenantId;
-  return typeof tenantId === "string" && tenantId.trim().length > 0
-    ? tenantId.trim()
-    : null;
-}
-
-function loadStoredLegacyCredentials(): AuthCredentials | null {
-  try {
-    const stored = sessionStorage.getItem(LEGACY_SESSION_AUTH_KEY);
-    if (!stored) return null;
-    // Migration credentials are one-shot: remove them from storage as soon as
-    // we read them, then keep them only in memory for the upgrade attempt.
-    sessionStorage.removeItem(LEGACY_SESSION_AUTH_KEY);
-
-    const parsed = JSON.parse(stored) as StoredLegacyAuthCredentials;
-    if (
-      !parsed ||
-      typeof parsed !== "object" ||
-      typeof parsed.username !== "string" ||
-      typeof parsed.password !== "string"
-    ) {
-      return null;
-    }
-
-    if (
-      typeof parsed.storedAt === "number" &&
-      Date.now() - parsed.storedAt > LEGACY_SESSION_AUTH_TTL_MS
-    ) {
-      return null;
-    }
-
-    return {
-      username: parsed.username,
-      password: parsed.password,
-    };
-  } catch {
-    return null;
-  }
-}
-
-function storeLegacyCredentials(credentials: AuthCredentials | null): void {
-  try {
-    if (credentials) {
-      sessionStorage.setItem(
-        LEGACY_SESSION_AUTH_KEY,
-        JSON.stringify({
-          ...credentials,
-          storedAt: Date.now(),
-        } satisfies StoredLegacyAuthCredentials),
-      );
-    } else {
-      sessionStorage.removeItem(LEGACY_SESSION_AUTH_KEY);
-    }
-  } catch {
-    // Ignore storage errors in restricted browser contexts.
-  }
-}
-
-function loadStoredAuthToken(): string | null {
-  try {
-    return (
-      sessionStorage.getItem(SESSION_AUTH_TOKEN_KEY) ??
-      sessionStorage.getItem(LEGACY_SESSION_JWT_KEY)
-    );
-  } catch {
-    return null;
-  }
-}
-
-function storeAuthToken(token: string | null): void {
-  try {
-    if (token) {
-      sessionStorage.setItem(SESSION_AUTH_TOKEN_KEY, token);
-      sessionStorage.removeItem(LEGACY_SESSION_JWT_KEY);
-    } else {
-      sessionStorage.removeItem(SESSION_AUTH_TOKEN_KEY);
-      sessionStorage.removeItem(LEGACY_SESSION_JWT_KEY);
-    }
-  } catch {
-    // Ignore storage errors in restricted browser contexts.
-  }
-}
-
-export function getCurrentAuthWorkspaceStorageScope(): string | null {
-  const tenantId = getTenantIdFromAuthToken(loadStoredAuthToken());
-  return tenantId ? `workspace:${tenantId}` : null;
-}
-
-export function getAuthScopedStorageKey(baseKey: string): string {
-  const scope = getCurrentAuthWorkspaceStorageScope();
-  return scope ? `${baseKey}:${scope}` : baseKey;
-}
-
-let cachedLegacyCredentials: AuthCredentials | null =
-  loadStoredLegacyCredentials();
-let cachedAuthToken: string | null = loadStoredAuthToken();
-let authMigrationInFlight: Promise<boolean> | null = null;
-
-function clearCachedAppData(): void {
-  queryClient.clear();
-}
-
-export function clearAuthSession(): void {
-  clearCachedAppData();
-  cachedLegacyCredentials = null;
-  cachedAuthToken = null;
-  storeLegacyCredentials(null);
-  storeAuthToken(null);
-}
-
-function setAuthenticatedSession(token: string): void {
-  clearCachedAppData();
-  cachedAuthToken = token;
-  storeAuthToken(token);
-  cachedLegacyCredentials = null;
-  storeLegacyCredentials(null);
-}
-
-async function readAuthResponse<T>(
-  response: Response,
-): Promise<ApiResponse<T> | LegacyApiResponse<T>> {
-  const text = await response.text();
-  let payload: unknown;
-  try {
-    payload = JSON.parse(text);
-  } catch {
-    throw new ApiClientError(
-      `Server error (${response.status}): Expected JSON but received HTML. Is the backend server running?`,
-      { status: response.status },
-    );
-  }
-
-  return normalizeApiResponse<T>(payload);
-}
-
-export async function signInWithCredentials(
-  username: string,
-  password: string,
-): Promise<void> {
-  const res = await fetch("/api/auth/login", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify({ username, password }),
-  });
-
-  const parsed = await readAuthResponse<{ token: string }>(res);
-  if ("ok" in parsed) {
-    if (!parsed.ok) {
-      throw toApiError(res, parsed);
-    }
-  } else if (!parsed.success) {
-    throw toApiError(res, parsed);
-  }
-
-  const token =
-    "ok" in parsed
-      ? parsed.data?.token
-      : (parsed.data as { token?: string } | undefined)?.token;
-  if (!token) {
-    throw new Error("No token returned");
-  }
-  setAuthenticatedSession(token);
-}
-
-export async function getAuthBootstrapStatus(): Promise<AuthBootstrapStatus> {
-  return fetchApi<AuthBootstrapStatus>("/auth/bootstrap-status");
-}
-
-export async function setupFirstAdmin(input: {
-  username: string;
-  password: string;
-  displayName?: string;
-}): Promise<AuthUser> {
-  const res = await fetch("/api/auth/setup", {
-    method: "POST",
-    headers: { "Content-Type": "application/json" },
-    body: JSON.stringify(input),
-  });
-  const parsed = await readAuthResponse<{
-    token: string;
-    user: AuthUser;
-  }>(res);
-  if ("ok" in parsed) {
-    if (!parsed.ok) throw toApiError(res, parsed);
-    if (!parsed.data?.token || !parsed.data.user) {
-      throw new Error("Setup response was incomplete");
-    }
-    setAuthenticatedSession(parsed.data.token);
-    return parsed.data.user;
-  }
-  if (!parsed.success) throw toApiError(res, parsed);
-  const data = parsed.data as { token?: string; user?: AuthUser } | undefined;
-  if (!data?.token || !data.user) {
-    throw new Error("Setup response was incomplete");
-  }
-  setAuthenticatedSession(data.token);
-  return data.user;
-}
-
-export async function getCurrentAuthUser(): Promise<AuthUser> {
-  const result = await fetchApi<{ user: AuthUser }>("/auth/me");
-  return result.user;
-}
-
-export async function restoreAuthSessionFromLegacyCredentials(): Promise<boolean> {
-  if (cachedAuthToken) return true;
-  if (!cachedLegacyCredentials) return false;
-  if (!authMigrationInFlight) {
-    const credentials = cachedLegacyCredentials;
-    cachedLegacyCredentials = null;
-    storeLegacyCredentials(null);
-    authMigrationInFlight = (async () => {
-      try {
-        await signInWithCredentials(credentials.username, credentials.password);
-        return true;
-      } catch {
-        return false;
-      } finally {
-        authMigrationInFlight = null;
-      }
-    })();
-  }
-  return authMigrationInFlight;
-}
-
-async function recoverAuthSessionFromUnauthorized(): Promise<string | null> {
-  cachedAuthToken = null;
-  storeAuthToken(null);
-
-  const restored = await restoreAuthSessionFromLegacyCredentials();
-  if (restored && cachedAuthToken) {
-    return `Bearer ${cachedAuthToken}`;
-  }
-
-  clearAuthSession();
-  redirectToSignIn();
-  return null;
-}
-
-export async function recoverAuthHeaderAfterUnauthorized(): Promise<
-  string | null
-> {
-  return recoverAuthSessionFromUnauthorized();
-}
-
-export async function logout(
-  options: { redirect?: boolean } = {},
-): Promise<void> {
-  if (cachedAuthToken) {
-    try {
-      await fetch("/api/auth/logout", {
-        method: "POST",
-        headers: { Authorization: `Bearer ${cachedAuthToken}` },
-      });
-    } catch {
-      // Best-effort server-side invalidation.
-    }
-  }
-  clearAuthSession();
-  if (options.redirect ?? true) {
-    redirectToSignIn();
-  }
-}
-
-export function getCachedAuthHeader(): string | undefined {
-  return cachedAuthToken ? `Bearer ${cachedAuthToken}` : undefined;
-}
-
-export function hasAuthenticatedSession(): boolean {
-  return Boolean(cachedAuthToken);
-}
-
-export async function listWorkspaceUsers(): Promise<AuthUser[]> {
-  const result = await fetchApi<{ users: AuthUser[] }>("/workspaces/users");
-  return result.users;
-}
-
-export async function createWorkspaceUser(input: {
-  username: string;
-  password: string;
-  displayName?: string;
-  isSystemAdmin?: boolean;
-}): Promise<AuthUser> {
-  const result = await fetchApi<{ user: AuthUser }>("/workspaces/users", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-  return result.user;
-}
-
-export async function setWorkspaceUserDisabled(
-  userId: string,
-  isDisabled: boolean,
-): Promise<AuthUser> {
-  const result = await fetchApi<{ user: AuthUser }>(
-    `/workspaces/users/${encodeURIComponent(userId)}/disabled`,
-    {
-      method: "PATCH",
-      body: JSON.stringify({ isDisabled }),
-    },
-  );
-  return result.user;
-}
-
-export async function resetWorkspaceUserPassword(
-  userId: string,
-  password: string,
-): Promise<void> {
-  await fetchApi<{ userId: string }>(
-    `/workspaces/users/${encodeURIComponent(userId)}/reset-password`,
-    {
-      method: "POST",
-      body: JSON.stringify({ password }),
-    },
-  );
-}
-
-export async function changeOwnPassword(password: string): Promise<void> {
-  await fetchApi<{ userId: string }>("/workspaces/me/password", {
-    method: "POST",
-    body: JSON.stringify({ password }),
-  });
-}
-
-export function __resetApiClientAuthForTests(): void {
-  cachedLegacyCredentials = null;
-  cachedAuthToken = null;
-  authMigrationInFlight = null;
-  storeLegacyCredentials(null);
-  storeAuthToken(null);
-}
-
-export function __setLegacyAuthCredentialsForTests(
-  credentials: AuthCredentials | null,
-): void {
-  cachedLegacyCredentials = credentials;
-  storeLegacyCredentials(credentials);
-}
-
-export function __setAuthTokenForTests(token: string | null): void {
-  cachedAuthToken = token;
-  storeAuthToken(token);
-}
-
-function normalizeApiResponse<T>(
-  payload: unknown,
-): ApiResponse<T> | LegacyApiResponse<T> {
-  if (!payload || typeof payload !== "object") {
-    throw new ApiClientError("API request failed: malformed JSON response");
-  }
-  const response = payload as Record<string, unknown>;
-  if (typeof response.ok === "boolean") {
-    return payload as ApiResponse<T>;
-  }
-  if (typeof response.success === "boolean") {
-    return payload as LegacyApiResponse<T>;
-  }
-  throw new ApiClientError("API request failed: unexpected response shape");
-}
-
-function describeAction(endpoint: string, method?: string): string {
-  const verb = (method || "GET").toUpperCase();
-  const normalized = endpoint.split("?")[0] || endpoint;
-  if (verb === "POST" && normalized === "/pipeline/run") {
-    return "Pipeline run used demo simulation.";
-  }
-  if (verb === "POST" && normalized.endsWith("/process")) {
-    return "Job processing used demo simulation.";
-  }
-  if (verb === "POST" && normalized.endsWith("/summarize")) {
-    return "Summary generation used demo simulation.";
-  }
-  if (verb === "POST" && normalized.endsWith("/generate-pdf")) {
-    return "PDF generation used demo simulation.";
-  }
-  if (verb === "POST" && normalized.endsWith("/rescore")) {
-    return "Suitability rescoring used demo simulation.";
-  }
-  if (verb === "POST" && normalized.endsWith("/apply")) {
-    return "Apply flow used demo simulation and no external sync.";
-  }
-  if (normalized.startsWith("/onboarding/validate")) {
-    return "Credential validation is simulated in demo mode.";
-  }
-  return "This action ran in demo simulation mode.";
-}
-
-function normalizeHeaders(headers?: HeadersInit): Record<string, string> {
-  if (!headers) return {};
-  if (headers instanceof Headers) {
-    const next: Record<string, string> = {};
-    headers.forEach((value, key) => {
-      next[key] = value;
-    });
-    return next;
-  }
-  if (Array.isArray(headers)) {
-    return Object.fromEntries(headers);
-  }
-  return { ...headers };
-}
-
-function isUnauthorizedResponse<T>(
-  response: Response,
-  parsed: ApiResponse<T> | LegacyApiResponse<T>,
-): boolean {
-  if (response.status !== 401) return false;
-  if ("ok" in parsed) {
-    return parsed.ok ? false : parsed.error.code === "UNAUTHORIZED";
-  }
-  return !parsed.success;
-}
-
-function toApiError<T>(
-  response: Response,
-  parsed: ApiResponse<T> | LegacyApiResponse<T>,
-): ApiClientError {
-  if ("ok" in parsed) {
-    if (!parsed.ok) {
-      return new ApiClientError(
-        formatUserFacingError(
-          {
-            message: parsed.error.message || "API request failed",
-            details: parsed.error.details,
-          },
-          "API request failed",
-        ),
-        {
-          requestId: parsed.meta?.requestId,
-          status: response.status,
-          code: parsed.error.code,
-        },
-      );
-    }
-    return new ApiClientError("API request failed", {
-      requestId: parsed.meta?.requestId,
-      status: response.status,
-    });
-  }
-  if (parsed.success) {
-    return new ApiClientError(
-      formatUserFacingError(parsed.message || "API request failed"),
-      {
-        status: response.status,
-      },
-    );
-  }
-  return new ApiClientError(
-    formatUserFacingError(parsed, "API request failed"),
-    {
-      status: response.status,
-    },
-  );
-}
-
-async function fetchAndParse<T>(
-  endpoint: string,
-  options: RequestInit | undefined,
-  authHeader?: string,
-): Promise<{
-  response: Response;
-  parsed: ApiResponse<T> | LegacyApiResponse<T>;
-}> {
-  const headers: Record<string, string> = {
-    "Content-Type": "application/json",
-    ...getAnalyticsRequestHeaders(),
-    ...normalizeHeaders(options?.headers),
-  };
-  if (authHeader) headers.Authorization = authHeader;
-  const response = await fetch(`${API_BASE}${endpoint}`, {
-    ...options,
-    headers,
-  });
-
-  const text = await response.text();
-
-  let payload: unknown;
-  try {
-    payload = JSON.parse(text);
-  } catch {
-    // If the response is not JSON, it's likely an HTML error page.
-    throw new ApiClientError(
-      `Server error (${response.status}): Expected JSON but received HTML. Is the backend server running?`,
-      { status: response.status },
-    );
-  }
-  const parsed = normalizeApiResponse<T>(payload);
-  return { response, parsed };
-}
-
-function getAuthHeader(): string | undefined {
-  return getCachedAuthHeader();
-}
-
-async function fetchApi<T>(
-  endpoint: string,
-  options?: RequestInit,
-): Promise<T> {
-  let authHeader = getAuthHeader();
-  let authAttempt = 0;
-
-  while (true) {
-    const { response, parsed } = await fetchAndParse(
-      endpoint,
-      options,
-      authHeader,
-    );
-
-    if (isUnauthorizedResponse(response, parsed) && authAttempt < 1) {
-      const recoveredAuthHeader = await recoverAuthSessionFromUnauthorized();
-      if (!recoveredAuthHeader) {
-        throw toApiError(response, parsed);
-      }
-      authHeader = recoveredAuthHeader;
-      authAttempt += 1;
-      continue;
-    }
-
-    if ("ok" in parsed) {
-      if (!parsed.ok) {
-        if (parsed.error.code === "UNAUTHORIZED") {
-          clearAuthSession();
-          redirectToSignIn();
-        }
-        if (parsed.meta?.blockedReason) {
-          showDemoBlockedToast(parsed.meta.blockedReason);
-        }
-        throw toApiError(response, parsed);
-      }
-      if (parsed.meta?.simulated) {
-        showDemoSimulatedToast(describeAction(endpoint, options?.method));
-      }
-      return parsed.data as T;
-    }
-
-    if (!parsed.success) {
-      if (response.status === 401) {
-        clearAuthSession();
-        redirectToSignIn();
-      }
-      throw toApiError(response, parsed);
-    }
-
-    const data = parsed.data;
-    if (data !== undefined) return data as T;
-    if (parsed.message !== undefined) return { message: parsed.message } as T;
-    return null as T;
-  }
-}
-
-async function fetchBlobApi(
-  endpoint: string,
-  options?: RequestInit,
-): Promise<Blob> {
-  let authHeader = getAuthHeader();
-  let authAttempt = 0;
-
-  while (true) {
-    const headers: Record<string, string> = {
-      ...getAnalyticsRequestHeaders(),
-      ...normalizeHeaders(options?.headers),
-    };
-    if (authHeader) headers.Authorization = authHeader;
-    const response = await fetch(`${API_BASE}${endpoint}`, {
-      ...options,
-      headers,
-    });
-
-    if (response.status === 401 && authAttempt < 1) {
-      const recoveredAuthHeader = await recoverAuthSessionFromUnauthorized();
-      if (recoveredAuthHeader) {
-        authHeader = recoveredAuthHeader;
-        authAttempt += 1;
-        continue;
-      }
-    }
-
-    if (!response.ok) {
-      const parsed = await readAuthResponse<never>(response);
-      throw toApiError(response, parsed);
-    }
-
-    return response.blob();
-  }
-}
-
-function normalizeApiPath(pathOrUrl: string): string {
-  const trimmed = pathOrUrl.trim();
-  if (!trimmed) {
-    return "/design-resume/pdf";
-  }
-
-  if (trimmed.startsWith("/api/")) {
-    return trimmed.slice(4);
-  }
-
-  if (trimmed.startsWith("/")) {
-    return trimmed;
-  }
-
-  return `/${trimmed}`;
-}
-
-// Jobs API
-export function getJobs(): Promise<JobsListResponse<JobListItem>>;
-export function getJobs(options: {
-  statuses?: string[];
-  view?: "list";
-}): Promise<JobsListResponse<JobListItem>>;
-export function getJobs(options?: {
-  statuses?: string[];
-  view: "full";
-}): Promise<JobsListResponse<Job>>;
-export async function getJobs(options?: {
-  statuses?: string[];
-  view?: "full" | "list";
-}): Promise<JobsListResponse<Job> | JobsListResponse<JobListItem>> {
-  const params = new URLSearchParams();
-  if (options?.statuses?.length)
-    params.set("status", options.statuses.join(","));
-  if (options?.view) params.set("view", options.view);
-  const query = params.toString();
-  return fetchApi<JobsListResponse<Job> | JobsListResponse<JobListItem>>(
-    `/jobs${query ? `?${query}` : ""}`,
-  );
-}
-
-export async function getJobsRevision(options?: {
-  statuses?: string[];
-}): Promise<JobsRevisionResponse> {
-  const params = new URLSearchParams();
-  if (options?.statuses?.length)
-    params.set("status", options.statuses.join(","));
-  const query = params.toString();
-  return fetchApi<JobsRevisionResponse>(
-    `/jobs/revision${query ? `?${query}` : ""}`,
-  );
-}
-
-export async function getJob(id: string): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}?t=${Date.now()}`);
-}
-
-export async function updateJob(
-  id: string,
-  update: Partial<Job>,
-): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}`, {
-    method: "PATCH",
-    body: JSON.stringify(update),
-  });
-}
-
-export async function getJobNotes(id: string): Promise<JobNote[]> {
-  return fetchApi<JobNote[]>(`/jobs/${id}/notes?t=${Date.now()}`);
-}
-
-export async function createJobNote(
-  jobId: string,
-  input: CreateJobNoteInput,
-): Promise<JobNote> {
-  return fetchApi<JobNote>(`/jobs/${jobId}/notes`, {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function updateJobNote(
-  jobId: string,
-  noteId: string,
-  input: UpdateJobNoteInput,
-): Promise<JobNote> {
-  return fetchApi<JobNote>(`/jobs/${jobId}/notes/${noteId}`, {
-    method: "PATCH",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function deleteJobNote(
-  jobId: string,
-  noteId: string,
-): Promise<void> {
-  await fetchApi<void>(`/jobs/${jobId}/notes/${noteId}`, {
-    method: "DELETE",
-  });
-}
-
-export async function uploadJobPdf(
-  id: string,
-  input: {
-    fileName: string;
-    mediaType?: string;
-    dataBase64: string;
-  },
-): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}/pdf`, {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function getJobPdfBlob(id: string): Promise<Blob> {
-  const cacheBuster = Date.now().toString(36);
-  return fetchBlobApi(`/jobs/${encodeURIComponent(id)}/pdf?v=${cacheBuster}`, {
-    cache: "no-store",
-  });
-}
-
-export async function getTracerAnalytics(options?: {
-  jobId?: string;
-  from?: number;
-  to?: number;
-  includeBots?: boolean;
-  limit?: number;
-}): Promise<TracerAnalyticsResponse> {
-  const params = new URLSearchParams();
-  if (options?.jobId) params.set("jobId", options.jobId);
-  if (typeof options?.from === "number") {
-    params.set("from", String(options.from));
-  }
-  if (typeof options?.to === "number") {
-    params.set("to", String(options.to));
-  }
-  if (typeof options?.includeBots === "boolean") {
-    params.set("includeBots", options.includeBots ? "1" : "0");
-  }
-  if (typeof options?.limit === "number") {
-    params.set("limit", String(options.limit));
-  }
-
-  const query = params.toString();
-  return fetchApi<TracerAnalyticsResponse>(
-    `/tracer-links/analytics${query ? `?${query}` : ""}`,
-  );
-}
-
-export async function getTracerReadiness(options?: {
-  force?: boolean;
-}): Promise<TracerReadinessResponse> {
-  const params = new URLSearchParams();
-  if (options?.force) params.set("force", "1");
-  const query = params.toString();
-  return fetchApi<TracerReadinessResponse>(
-    `/tracer-links/readiness${query ? `?${query}` : ""}`,
-  );
-}
-
-export async function getJobTracerLinks(
-  jobId: string,
-  options?: {
-    from?: number;
-    to?: number;
-    includeBots?: boolean;
-  },
-): Promise<JobTracerLinksResponse> {
-  const params = new URLSearchParams();
-  if (typeof options?.from === "number") {
-    params.set("from", String(options.from));
-  }
-  if (typeof options?.to === "number") {
-    params.set("to", String(options.to));
-  }
-  if (typeof options?.includeBots === "boolean") {
-    params.set("includeBots", options.includeBots ? "1" : "0");
-  }
-  const query = params.toString();
-  return fetchApi<JobTracerLinksResponse>(
-    `/tracer-links/jobs/${encodeURIComponent(jobId)}${query ? `?${query}` : ""}`,
-  );
-}
-
-async function streamSseEvents<TEvent>(
-  endpoint: string,
-  input: StreamSseInput,
-  handlers: {
-    onEvent: (event: TEvent) => void;
-    signal?: AbortSignal;
-  },
-): Promise<void> {
-  const headers: Record<string, string> = {
-    "Content-Type": "application/json",
-    ...getAnalyticsRequestHeaders(),
-  };
-  const streamAuth = getAuthHeader();
-  if (streamAuth) {
-    headers.Authorization = streamAuth;
-  }
-
-  let response = await fetch(`${API_BASE}${endpoint}`, {
-    method: "POST",
-    headers,
-    body: JSON.stringify(input),
-    signal: handlers.signal,
-  });
-
-  if (response.status === 401) {
-    const recoveredAuthHeader = await recoverAuthSessionFromUnauthorized();
-    if (recoveredAuthHeader) {
-      response = await fetch(`${API_BASE}${endpoint}`, {
-        method: "POST",
-        headers: {
-          ...headers,
-          Authorization: recoveredAuthHeader,
-        },
-        body: JSON.stringify(input),
-        signal: handlers.signal,
-      });
-    }
-  }
-
-  if (!response.ok) {
-    let errorMessage = `Stream request failed with status ${response.status}`;
-    try {
-      const payload = await response.json();
-      const parsed = normalizeApiResponse(payload);
-      if ("ok" in parsed && !parsed.ok) {
-        errorMessage = formatUserFacingError(
-          {
-            message: parsed.error.message || errorMessage,
-            details: parsed.error.details,
-          },
-          errorMessage,
-        );
-      }
-    } catch {
-      // ignore parse errors; keep status-based message
-    }
-    throw new ApiClientError(errorMessage, {
-      status: response.status,
-    });
-  }
-
-  if (!response.body) {
-    throw new ApiClientError("Streaming not supported by this browser");
-  }
-
-  const decoder = new TextDecoder();
-  const reader = response.body.getReader();
-  let buffer = "";
-
-  try {
-    while (true) {
-      const { done, value } = await reader.read();
-      if (done) break;
-      buffer += decoder.decode(value, { stream: true });
-
-      let separatorIndex = buffer.indexOf("\n\n");
-      while (separatorIndex !== -1) {
-        const frame = buffer.slice(0, separatorIndex);
-        buffer = buffer.slice(separatorIndex + 2);
-        const dataLines = frame
-          .split("\n")
-          .filter((line) => line.startsWith("data:"))
-          .map((line) => line.slice(5).trim())
-          .filter(Boolean);
-
-        for (const line of dataLines) {
-          let parsedEvent: TEvent;
-          try {
-            parsedEvent = JSON.parse(line) as TEvent;
-          } catch {
-            // Ignore malformed events to keep stream resilient
-            continue;
-          }
-          handlers.onEvent(parsedEvent);
-        }
-        separatorIndex = buffer.indexOf("\n\n");
-      }
-    }
-  } finally {
-    try {
-      await reader.cancel();
-    } catch {
-      // Ignore cancellation errors when stream is already closed
-    }
-  }
-}
-
-export async function listJobChatThreads(jobId: string): Promise<{
-  threads: JobChatThread[];
-}> {
-  return fetchApi<{ threads: JobChatThread[] }>(`/jobs/${jobId}/chat/threads`);
-}
-
-export async function listJobGhostwriterMessages(
-  jobId: string,
-  options?: { limit?: number; offset?: number },
-): Promise<{
-  messages: JobChatMessage[];
-  branches: BranchInfo[];
-  selectedNoteIds: string[];
-}> {
-  const params = new URLSearchParams();
-  if (typeof options?.limit === "number") {
-    params.set("limit", String(options.limit));
-  }
-  if (typeof options?.offset === "number") {
-    params.set("offset", String(options.offset));
-  }
-  const query = params.toString();
-  return fetchApi<{
-    messages: JobChatMessage[];
-    branches: BranchInfo[];
-    selectedNoteIds: string[];
-  }>(`/jobs/${jobId}/chat/messages${query ? `?${query}` : ""}`);
-}
-
-export async function updateJobGhostwriterContext(
-  jobId: string,
-  input: { selectedNoteIds: string[] },
-): Promise<{ selectedNoteIds: string[] }> {
-  return fetchApi<{ selectedNoteIds: string[] }>(
-    `/jobs/${jobId}/chat/context`,
-    {
-      method: "PATCH",
-      body: JSON.stringify(input),
-    },
-  );
-}
-
-export async function createJobChatThread(
-  jobId: string,
-  input?: { title?: string | null },
-): Promise<{ thread: JobChatThread }> {
-  return fetchApi<{ thread: JobChatThread }>(`/jobs/${jobId}/chat/threads`, {
-    method: "POST",
-    body: JSON.stringify({
-      title: input?.title ?? null,
-    }),
-  });
-}
-
-export async function listJobChatMessages(
-  jobId: string,
-  threadId: string,
-  options?: { limit?: number; offset?: number },
-): Promise<{ messages: JobChatMessage[] }> {
-  const params = new URLSearchParams();
-  if (typeof options?.limit === "number") {
-    params.set("limit", String(options.limit));
-  }
-  if (typeof options?.offset === "number") {
-    params.set("offset", String(options.offset));
-  }
-  const query = params.toString();
-  return fetchApi<{ messages: JobChatMessage[] }>(
-    `/jobs/${jobId}/chat/threads/${threadId}/messages${query ? `?${query}` : ""}`,
-  );
-}
-
-export async function sendJobChatMessage(
-  jobId: string,
-  threadId: string,
-  input: { content: string; selectedNoteIds?: string[] },
-): Promise<{
-  userMessage: JobChatMessage;
-  assistantMessage: JobChatMessage | null;
-  runId: string;
-}> {
-  return fetchApi<{
-    userMessage: JobChatMessage;
-    assistantMessage: JobChatMessage | null;
-    runId: string;
-  }>(`/jobs/${jobId}/chat/threads/${threadId}/messages`, {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function streamJobChatMessage(
-  jobId: string,
-  threadId: string,
-  input: { content: string; selectedNoteIds?: string[]; signal?: AbortSignal },
-  handlers: {
-    onEvent: (event: JobChatStreamEvent) => void;
-  },
-): Promise<void> {
-  return streamSseEvents(
-    `/jobs/${jobId}/chat/threads/${threadId}/messages`,
-    {
-      content: input.content,
-      selectedNoteIds: input.selectedNoteIds,
-      stream: true,
-    },
-    {
-      onEvent: handlers.onEvent,
-      signal: input.signal,
-    },
-  );
-}
-
-export async function streamJobGhostwriterMessage(
-  jobId: string,
-  input: { content: string; selectedNoteIds?: string[]; signal?: AbortSignal },
-  handlers: {
-    onEvent: (event: JobChatStreamEvent) => void;
-  },
-): Promise<void> {
-  return streamSseEvents(
-    `/jobs/${jobId}/chat/messages`,
-    {
-      content: input.content,
-      selectedNoteIds: input.selectedNoteIds,
-      stream: true,
-    },
-    {
-      onEvent: handlers.onEvent,
-      signal: input.signal,
-    },
-  );
-}
-
-export async function cancelJobChatRun(
-  jobId: string,
-  threadId: string,
-  runId: string,
-): Promise<{ cancelled: boolean; alreadyFinished: boolean }> {
-  return fetchApi<{ cancelled: boolean; alreadyFinished: boolean }>(
-    `/jobs/${jobId}/chat/threads/${threadId}/runs/${runId}/cancel`,
-    {
-      method: "POST",
-      body: JSON.stringify({}),
-    },
-  );
-}
-
-export async function resetJobGhostwriterConversation(
-  jobId: string,
-): Promise<{ deletedMessages: number; deletedRuns: number }> {
-  return fetchApi<{ deletedMessages: number; deletedRuns: number }>(
-    `/jobs/${jobId}/chat/reset`,
-    {
-      method: "POST",
-      body: JSON.stringify({}),
-    },
-  );
-}
-
-export async function cancelJobGhostwriterRun(
-  jobId: string,
-  runId: string,
-): Promise<{ cancelled: boolean; alreadyFinished: boolean }> {
-  return fetchApi<{ cancelled: boolean; alreadyFinished: boolean }>(
-    `/jobs/${jobId}/chat/runs/${runId}/cancel`,
-    {
-      method: "POST",
-      body: JSON.stringify({}),
-    },
-  );
-}
-
-export async function regenerateJobChatMessage(
-  jobId: string,
-  threadId: string,
-  assistantMessageId: string,
-): Promise<{ runId: string; assistantMessage: JobChatMessage | null }> {
-  return fetchApi<{ runId: string; assistantMessage: JobChatMessage | null }>(
-    `/jobs/${jobId}/chat/threads/${threadId}/messages/${assistantMessageId}/regenerate`,
-    {
-      method: "POST",
-      body: JSON.stringify({}),
-    },
-  );
-}
-
-export async function streamRegenerateJobChatMessage(
-  jobId: string,
-  threadId: string,
-  assistantMessageId: string,
-  input: { selectedNoteIds?: string[]; signal?: AbortSignal },
-  handlers: {
-    onEvent: (event: JobChatStreamEvent) => void;
-  },
-): Promise<void> {
-  return streamSseEvents(
-    `/jobs/${jobId}/chat/threads/${threadId}/messages/${assistantMessageId}/regenerate`,
-    { selectedNoteIds: input.selectedNoteIds, stream: true },
-    {
-      onEvent: handlers.onEvent,
-      signal: input.signal,
-    },
-  );
-}
-
-export async function streamRegenerateJobGhostwriterMessage(
-  jobId: string,
-  assistantMessageId: string,
-  input: { selectedNoteIds?: string[]; signal?: AbortSignal },
-  handlers: {
-    onEvent: (event: JobChatStreamEvent) => void;
-  },
-): Promise<void> {
-  return streamSseEvents(
-    `/jobs/${jobId}/chat/messages/${assistantMessageId}/regenerate`,
-    { selectedNoteIds: input.selectedNoteIds, stream: true },
-    {
-      onEvent: handlers.onEvent,
-      signal: input.signal,
-    },
-  );
-}
-
-export async function editJobGhostwriterMessage(
-  jobId: string,
-  messageId: string,
-  input: { content: string; selectedNoteIds?: string[]; signal?: AbortSignal },
-  handlers: {
-    onEvent: (event: JobChatStreamEvent) => void;
-  },
-): Promise<void> {
-  return streamSseEvents(
-    `/jobs/${jobId}/chat/messages/${messageId}/edit`,
-    {
-      content: input.content,
-      selectedNoteIds: input.selectedNoteIds,
-      stream: true,
-    },
-    {
-      onEvent: handlers.onEvent,
-      signal: input.signal,
-    },
-  );
-}
-
-export async function switchJobGhostwriterBranch(
-  jobId: string,
-  messageId: string,
-): Promise<{ messages: JobChatMessage[]; branches: BranchInfo[] }> {
-  return fetchApi<{ messages: JobChatMessage[]; branches: BranchInfo[] }>(
-    `/jobs/${jobId}/chat/messages/${messageId}/switch-branch`,
-    {
-      method: "POST",
-      body: JSON.stringify({}),
-    },
-  );
-}
-
-function toJobIdList(idOrIds: string | string[]): string[] {
-  return Array.isArray(idOrIds) ? idOrIds : [idOrIds];
-}
-
-export async function processJob(
-  ids: string[],
-  options?: { force?: boolean },
-): Promise<JobActionResponse>;
-export async function processJob(
-  id: string,
-  options?: { force?: boolean },
-): Promise<Job>;
-export async function processJob(
-  idOrIds: string | string[],
-  options?: { force?: boolean },
-): Promise<Job | JobActionResponse> {
-  const jobIds = toJobIdList(idOrIds);
-  const result = await runJobAction({
-    action: "move_to_ready",
-    jobIds,
-    ...(options?.force ? { options: { force: true } } : {}),
-  });
-
-  if (Array.isArray(idOrIds)) return result;
-  return getSingleJobFromActionResult(result, idOrIds);
-}
-
-export async function rescoreJob(ids: string[]): Promise<JobActionResponse>;
-export async function rescoreJob(id: string): Promise<Job>;
-export async function rescoreJob(
-  idOrIds: string | string[],
-): Promise<Job | JobActionResponse> {
-  const jobIds = toJobIdList(idOrIds);
-  const result = await runJobAction({
-    action: "rescore",
-    jobIds,
-  });
-  if (Array.isArray(idOrIds)) return result;
-  return getSingleJobFromActionResult(result, idOrIds);
-}
-
-export async function summarizeJob(
-  id: string,
-  options?: { force?: boolean },
-): Promise<Job> {
-  const query = options?.force ? "?force=1" : "";
-  return fetchApi<Job>(`/jobs/${id}/summarize${query}`, {
-    method: "POST",
-  });
-}
-
-export async function generateJobPdf(id: string): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}/generate-pdf`, {
-    method: "POST",
-  });
-}
-
-export async function checkSponsor(id: string): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}/check-sponsor`, {
-    method: "POST",
-  });
-}
-
-export async function markAsApplied(id: string): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}/apply`, {
-    method: "POST",
-  });
-}
-
-export async function skipJob(ids: string[]): Promise<JobActionResponse>;
-export async function skipJob(id: string): Promise<Job>;
-export async function skipJob(
-  idOrIds: string | string[],
-): Promise<Job | JobActionResponse> {
-  const jobIds = toJobIdList(idOrIds);
-  const result = await runJobAction({
-    action: "skip",
-    jobIds,
-  });
-  if (Array.isArray(idOrIds)) return result;
-  return getSingleJobFromActionResult(result, idOrIds);
-}
-
-export async function runJobAction(
-  input: JobActionRequest,
-): Promise<JobActionResponse> {
-  return fetchApi<JobActionResponse>("/jobs/actions", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-function getSingleJobFromActionResult(
-  response: JobActionResponse,
-  jobId: string,
-): Job {
-  const result = response.results.find((entry) => entry.jobId === jobId);
-  if (!result) {
-    throw new ApiClientError("Job action did not return a result for the job");
-  }
-  if (!result.ok) {
-    throw new ApiClientError(formatUserFacingError(result.error.message), {
-      code: result.error.code,
-    });
-  }
-  return result.job;
-}
-
-export async function streamJobAction(
-  input: JobActionRequest,
-  handlers: {
-    onEvent: (event: JobActionStreamEvent) => void;
-    signal?: AbortSignal;
-  },
-): Promise<void> {
-  return streamSseEvents<JobActionStreamEvent>(
-    "/jobs/actions/stream",
-    input,
-    handlers,
-  );
-}
-
-export async function getJobStageEvents(id: string): Promise<StageEvent[]> {
-  return fetchApi<StageEvent[]>(`/jobs/${id}/events?t=${Date.now()}`);
-}
-
-export async function getJobTasks(
-  id: string,
-  options?: { includeCompleted?: boolean },
-): Promise<ApplicationTask[]> {
-  const params = new URLSearchParams();
-  if (options?.includeCompleted) params.set("includeCompleted", "1");
-  params.set("t", Date.now().toString());
-  const query = params.toString();
-  return fetchApi<ApplicationTask[]>(`/jobs/${id}/tasks?${query}`);
-}
-
-export async function transitionJobStage(
-  id: string,
-  input: {
-    toStage: StageTransitionTarget;
-    occurredAt?: number | null;
-    metadata?: StageEventMetadata | null;
-    outcome?: JobOutcome | null;
-  },
-): Promise<StageEvent> {
-  return fetchApi<StageEvent>(`/jobs/${id}/stages`, {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function updateJobStageEvent(
-  id: string,
-  eventId: string,
-  input: {
-    toStage?: ApplicationStage;
-    occurredAt?: number | null;
-    metadata?: StageEventMetadata | null;
-    outcome?: JobOutcome | null;
-  },
-): Promise<void> {
-  return fetchApi<void>(`/jobs/${id}/events/${eventId}`, {
-    method: "PATCH",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function deleteJobStageEvent(
-  id: string,
-  eventId: string,
-): Promise<void> {
-  return fetchApi<void>(`/jobs/${id}/events/${eventId}`, {
-    method: "DELETE",
-  });
-}
-
-export async function updateJobOutcome(
-  id: string,
-  input: { outcome: JobOutcome | null; closedAt?: number | null },
-): Promise<Job> {
-  return fetchApi<Job>(`/jobs/${id}/outcome`, {
-    method: "PATCH",
-    body: JSON.stringify(input),
-  });
-}
-
-// Pipeline API
-export async function getPipelineStatus(): Promise<PipelineStatusResponse> {
-  return fetchApi<PipelineStatusResponse>("/pipeline/status");
-}
-
-export async function getPipelineProgressSnapshot(): Promise<PipelineProgressState> {
-  return fetchApi<PipelineProgressState>("/pipeline/progress/snapshot");
-}
-
-export async function getPipelineRuns(): Promise<PipelineRun[]> {
-  return fetchApi<PipelineRun[]>("/pipeline/runs");
-}
-
-export async function prepareChallengeViewer(): Promise<{
-  available: boolean;
-  viewerUrl: string | null;
-  reason: string | null;
-}> {
-  return fetchApi<{
-    available: boolean;
-    viewerUrl: string | null;
-    reason: string | null;
-  }>("/pipeline/challenge-viewer", {
-    method: "POST",
-  });
-}
-
-export async function solvePipelineChallenge(extractorId: string): Promise<{
-  status: "solved";
-  extractorId: string;
-  challengesRemaining: number;
-}> {
-  return fetchApi<{
-    status: "solved";
-    extractorId: string;
-    challengesRemaining: number;
-  }>("/pipeline/solve-challenge", {
-    method: "POST",
-    body: JSON.stringify({ extractorId }),
-  });
-}
-
-export async function getPipelineRunInsights(
-  id: string,
-): Promise<PipelineRunInsights> {
-  return fetchApi<PipelineRunInsights>(
-    `/pipeline/runs/${encodeURIComponent(id)}/insights`,
-  );
-}
-
-export async function runPipeline(config?: {
-  topN?: number;
-  minSuitabilityScore?: number;
-  sources?: JobSource[];
-  runBudget?: number;
-  searchTerms?: string[];
-  country?: string;
-  cityLocations?: string[];
-  workplaceTypes?: Array<"remote" | "hybrid" | "onsite">;
-  searchScope?: LocationSearchScope;
-  matchStrictness?: LocationMatchStrictness;
-}): Promise<{ message: string }> {
-  return fetchApi<{ message: string }>("/pipeline/run", {
-    method: "POST",
-    body: JSON.stringify(config || {}),
-  });
-}
-
-export async function cancelPipeline(): Promise<{
-  message: string;
-  pipelineRunId: string | null;
-  alreadyRequested: boolean;
-}> {
-  return fetchApi<{
-    message: string;
-    pipelineRunId: string | null;
-    alreadyRequested: boolean;
-  }>("/pipeline/cancel", {
-    method: "POST",
-  });
-}
-
-// Post-Application Tracking API
-export async function postApplicationProviderConnect(input: {
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  payload?: Record<string, unknown>;
-}): Promise<PostApplicationProviderActionResponse> {
-  const provider = input.provider ?? "gmail";
-  return fetchApi<PostApplicationProviderActionResponse>(
-    `/post-application/providers/${provider}/actions/connect`,
-    {
-      method: "POST",
-      body: JSON.stringify({
-        ...(input.accountKey ? { accountKey: input.accountKey } : {}),
-        ...(input.payload ? { payload: input.payload } : {}),
-      }),
-    },
-  );
-}
-
-export async function postApplicationGmailOauthStart(input?: {
-  accountKey?: string;
-}): Promise<{
-  provider: "gmail";
-  accountKey: string;
-  authorizationUrl: string;
-  state: string;
-}> {
-  const params = new URLSearchParams();
-  if (input?.accountKey) params.set("accountKey", input.accountKey);
-  const query = params.toString();
-  return fetchApi<{
-    provider: "gmail";
-    accountKey: string;
-    authorizationUrl: string;
-    state: string;
-  }>(
-    `/post-application/providers/gmail/oauth/start${query ? `?${query}` : ""}`,
-  );
-}
-
-export async function postApplicationGmailOauthExchange(input: {
-  accountKey?: string;
-  state: string;
-  code: string;
-}): Promise<PostApplicationProviderActionResponse> {
-  return fetchApi<PostApplicationProviderActionResponse>(
-    "/post-application/providers/gmail/oauth/exchange",
-    {
-      method: "POST",
-      body: JSON.stringify({
-        ...(input.accountKey ? { accountKey: input.accountKey } : {}),
-        state: input.state,
-        code: input.code,
-      }),
-    },
-  );
-}
-
-export async function postApplicationProviderStatus(input?: {
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-}): Promise<PostApplicationProviderActionResponse> {
-  const provider = input?.provider ?? "gmail";
-  return fetchApi<PostApplicationProviderActionResponse>(
-    `/post-application/providers/${provider}/actions/status`,
-    {
-      method: "POST",
-      body: JSON.stringify({
-        ...(input?.accountKey ? { accountKey: input.accountKey } : {}),
-      }),
-    },
-  );
-}
-
-export async function postApplicationProviderSync(input?: {
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  maxMessages?: number;
-  searchDays?: number;
-}): Promise<PostApplicationProviderActionResponse> {
-  const provider = input?.provider ?? "gmail";
-  return fetchApi<PostApplicationProviderActionResponse>(
-    `/post-application/providers/${provider}/actions/sync`,
-    {
-      method: "POST",
-      body: JSON.stringify({
-        ...(input?.accountKey ? { accountKey: input.accountKey } : {}),
-        ...(typeof input?.maxMessages === "number"
-          ? { maxMessages: input.maxMessages }
-          : {}),
-        ...(typeof input?.searchDays === "number"
-          ? { searchDays: input.searchDays }
-          : {}),
-      }),
-    },
-  );
-}
-
-export async function postApplicationProviderDisconnect(input?: {
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-}): Promise<PostApplicationProviderActionResponse> {
-  const provider = input?.provider ?? "gmail";
-  return fetchApi<PostApplicationProviderActionResponse>(
-    `/post-application/providers/${provider}/actions/disconnect`,
-    {
-      method: "POST",
-      body: JSON.stringify({
-        ...(input?.accountKey ? { accountKey: input.accountKey } : {}),
-      }),
-    },
-  );
-}
-
-export async function getPostApplicationInbox(input?: {
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  limit?: number;
-}): Promise<{ items: PostApplicationInboxItem[]; total: number }> {
-  const params = new URLSearchParams();
-  params.set("provider", input?.provider ?? "gmail");
-  params.set("accountKey", input?.accountKey ?? "default");
-  if (typeof input?.limit === "number")
-    params.set("limit", String(input.limit));
-  const query = params.toString();
-  return fetchApi<{ items: PostApplicationInboxItem[]; total: number }>(
-    `/post-application/inbox?${query}`,
-  );
-}
-
-export async function approvePostApplicationInboxItem(input: {
-  messageId: string;
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  jobId?: string;
-  stageTarget?: PostApplicationRouterStageTarget;
-  toStage?: ApplicationStage;
-  note?: string;
-  decidedBy?: string;
-}): Promise<{
-  message: PostApplicationInboxItem["message"];
-  stageEventId: string | null;
-}> {
-  return fetchApi<{
-    message: PostApplicationInboxItem["message"];
-    stageEventId: string | null;
-  }>(`/post-application/inbox/${encodeURIComponent(input.messageId)}/approve`, {
-    method: "POST",
-    body: JSON.stringify({
-      provider: input.provider ?? "gmail",
-      accountKey: input.accountKey ?? "default",
-      ...(input.jobId ? { jobId: input.jobId } : {}),
-      ...(input.stageTarget ? { stageTarget: input.stageTarget } : {}),
-      ...(input.toStage ? { toStage: input.toStage } : {}),
-      ...(input.note ? { note: input.note } : {}),
-      ...(input.decidedBy ? { decidedBy: input.decidedBy } : {}),
-    }),
-  });
-}
-
-export async function denyPostApplicationInboxItem(input: {
-  messageId: string;
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  decidedBy?: string;
-}): Promise<{
-  message: PostApplicationInboxItem["message"];
-}> {
-  return fetchApi<{ message: PostApplicationInboxItem["message"] }>(
-    `/post-application/inbox/${encodeURIComponent(input.messageId)}/deny`,
-    {
-      method: "POST",
-      body: JSON.stringify({
-        provider: input.provider ?? "gmail",
-        accountKey: input.accountKey ?? "default",
-        ...(input.decidedBy ? { decidedBy: input.decidedBy } : {}),
-      }),
-    },
-  );
-}
-
-export async function runPostApplicationInboxAction(input: {
-  action: PostApplicationAction;
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  decidedBy?: string;
-}): Promise<PostApplicationActionResponse> {
-  return fetchApi<PostApplicationActionResponse>(
-    "/post-application/inbox/actions",
-    {
-      method: "POST",
-      body: JSON.stringify({
-        action: input.action,
-        provider: input.provider ?? "gmail",
-        accountKey: input.accountKey ?? "default",
-        ...(input.decidedBy ? { decidedBy: input.decidedBy } : {}),
-      }),
-    },
-  );
-}
-
-export async function getPostApplicationRuns(input?: {
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  limit?: number;
-}): Promise<{ runs: PostApplicationSyncRun[]; total: number }> {
-  const params = new URLSearchParams();
-  params.set("provider", input?.provider ?? "gmail");
-  params.set("accountKey", input?.accountKey ?? "default");
-  if (typeof input?.limit === "number")
-    params.set("limit", String(input.limit));
-  const query = params.toString();
-  return fetchApi<{ runs: PostApplicationSyncRun[]; total: number }>(
-    `/post-application/runs?${query}`,
-  );
-}
-
-export async function getPostApplicationRunMessages(input: {
-  runId: string;
-  provider?: PostApplicationProvider;
-  accountKey?: string;
-  limit?: number;
-}): Promise<{
-  run: PostApplicationSyncRun;
-  items: PostApplicationInboxItem[];
-  total: number;
-}> {
-  const params = new URLSearchParams();
-  params.set("provider", input.provider ?? "gmail");
-  params.set("accountKey", input.accountKey ?? "default");
-  if (typeof input.limit === "number") params.set("limit", String(input.limit));
-  const query = params.toString();
-  return fetchApi<{
-    run: PostApplicationSyncRun;
-    items: PostApplicationInboxItem[];
-    total: number;
-  }>(
-    `/post-application/runs/${encodeURIComponent(input.runId)}/messages?${query}`,
-  );
-}
-
-export async function getDemoInfo(): Promise<DemoInfoResponse> {
-  return fetchApi<DemoInfoResponse>("/demo/info");
-}
-
-// Manual Job Import API
-export async function fetchJobFromUrl(input: {
-  url: string;
-}): Promise<ManualJobFetchResponse> {
-  return fetchApi<ManualJobFetchResponse>("/manual-jobs/fetch", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function inferManualJob(input: {
-  jobDescription: string;
-}): Promise<ManualJobInferenceResponse> {
-  return fetchApi<ManualJobInferenceResponse>("/manual-jobs/infer", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function importManualJob(input: {
-  job: ManualJobDraft;
-}): Promise<Job> {
-  return fetchApi<Job>("/manual-jobs/import", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-// Settings & Profile API
-let settingsPromise: Promise<AppSettings> | null = null;
-
-export async function getSettings(): Promise<AppSettings> {
-  if (settingsPromise) return settingsPromise;
-
-  settingsPromise = fetchApi<AppSettings>("/settings").finally(() => {
-    // Clear the promise after a short delay to allow subsequent fresh fetches
-    // but coalesce simultaneous requests.
-    setTimeout(() => {
-      settingsPromise = null;
-    }, 100);
-  });
-
-  return settingsPromise;
-}
-
-export async function getProfileProjects(): Promise<
-  ResumeProjectCatalogItem[]
-> {
-  return fetchApi<ResumeProjectCatalogItem[]>("/profile/projects");
-}
-
-export async function getResumeProjectsCatalog(): Promise<
-  ResumeProjectCatalogItem[]
-> {
-  // Always resolve from /profile/projects so local Design Resume edits
-  // propagate to active and future application tailoring flows.
-  return getProfileProjects();
-}
-
-export async function getProfile(): Promise<ResumeProfile> {
-  return fetchApi<ResumeProfile>("/profile");
-}
-
-export async function getDesignResume(): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume");
-}
-
-export async function getDesignResumeStatus(): Promise<DesignResumeStatusResponse> {
-  return fetchApi<DesignResumeStatusResponse>("/design-resume/status");
-}
-
-export async function importDesignResumeFromRxResume(): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume/import/rxresume", {
-    method: "POST",
-  });
-}
-
-export async function importDesignResumeFromFile(input: {
-  fileName: string;
-  mediaType?: string;
-  dataBase64: string;
-}): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume/import/file", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function updateDesignResume(
-  input: DesignResumePatchRequest,
-): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume", {
-    method: "PATCH",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function uploadDesignResumePicture(input: {
-  fileName: string;
-  dataUrl: string;
-  baseRevision?: number;
-  document?: DesignResumeJson;
-}): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume/assets", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function uploadDesignResumePictureFile(input: {
-  file: File;
-  baseRevision?: number;
-}): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume/assets", {
-    method: "POST",
-    headers: {
-      "Content-Type": input.file.type || "application/octet-stream",
-      "x-file-name": encodeURIComponent(input.file.name || "picture"),
-      ...(input.baseRevision
-        ? { "x-base-revision": String(input.baseRevision) }
-        : {}),
-    },
-    body: await input.file.arrayBuffer(),
-  });
-}
-
-export async function deleteDesignResumePicture(input?: {
-  baseRevision?: number;
-  document?: DesignResumeJson;
-}): Promise<DesignResumeDocument> {
-  return fetchApi<DesignResumeDocument>("/design-resume/assets/picture", {
-    method: "DELETE",
-    body: JSON.stringify(input ?? {}),
-  });
-}
-
-export async function exportDesignResume(): Promise<DesignResumeExportResponse> {
-  return fetchApi<DesignResumeExportResponse>("/design-resume/export");
-}
-
-export async function generateDesignResumePdf(): Promise<DesignResumePdfResponse> {
-  return fetchApi<DesignResumePdfResponse>("/design-resume/generate-pdf", {
-    method: "POST",
-  });
-}
-
-export async function getDesignResumePdfBlob(pdfUrl?: string): Promise<Blob> {
-  return fetchBlobApi(
-    pdfUrl ? normalizeApiPath(pdfUrl) : "/design-resume/pdf",
-    { cache: "no-store" },
-  );
-}
-
-export async function getProfileStatus(): Promise<ProfileStatusResponse> {
-  return fetchApi<ProfileStatusResponse>("/profile/status");
-}
-
-export async function refreshProfile(): Promise<ResumeProfile> {
-  return fetchApi<ResumeProfile>("/profile/refresh", {
-    method: "POST",
-  });
-}
-
-export async function validateLlm(input: {
-  provider?: string;
-  baseUrl?: string;
-  apiKey?: string;
-}): Promise<ValidationResult> {
-  return fetchApi<ValidationResult>("/onboarding/validate/llm", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function getLlmModels(input?: {
-  provider?: string;
-  baseUrl?: string;
-  apiKey?: string;
-}): Promise<string[]> {
-  const data = await fetchApi<{ models: string[] }>("/settings/llm-models", {
-    method: "POST",
-    body: JSON.stringify(input ?? {}),
-  });
-  return data.models;
-}
-
-export async function getCodexAuthStatus(): Promise<CodexAuthStatusResponse> {
-  return fetchApi<CodexAuthStatusResponse>("/settings/codex-auth");
-}
-
-export async function startCodexAuth(input?: {
-  forceRestart?: boolean;
-}): Promise<CodexAuthStatusResponse> {
-  return fetchApi<CodexAuthStatusResponse>("/settings/codex-auth/start", {
-    method: "POST",
-    body: JSON.stringify({
-      forceRestart: input?.forceRestart ?? false,
-    }),
-  });
-}
-
-export async function disconnectCodexAuth(): Promise<CodexAuthStatusResponse> {
-  return fetchApi<CodexAuthStatusResponse>("/settings/codex-auth/disconnect", {
-    method: "POST",
-  });
-}
-
-export async function validateRxresume(input?: {
-  apiKey?: string;
-  baseUrl?: string;
-}): Promise<ValidationResult> {
-  return fetchApi<ValidationResult>("/onboarding/validate/rxresume", {
-    method: "POST",
-    body: JSON.stringify(input ?? {}),
-  });
-}
-
-export async function validateResumeConfig(): Promise<ValidationResult> {
-  return fetchApi<ValidationResult>("/onboarding/validate/resume");
-}
-
-export async function suggestOnboardingSearchTerms(): Promise<SearchTermsSuggestionResponse> {
-  return fetchApi<SearchTermsSuggestionResponse>(
-    "/onboarding/search-terms/suggest",
-    {
-      method: "POST",
-    },
-  );
-}
-
-export async function updateSettings(
-  update: Partial<UpdateSettingsInput>,
-): Promise<AppSettings> {
-  return fetchApi<AppSettings>("/settings", {
-    method: "PATCH",
-    body: JSON.stringify(update),
-  });
-}
-
-export async function getRxResumes(): Promise<{ id: string; name: string }[]> {
-  const data = await fetchApi<{ resumes: { id: string; name: string }[] }>(
-    `/settings/rx-resumes`,
-  );
-  return data.resumes;
-}
-
-export async function getRxResumeProjects(
-  resumeId: string,
-  signal?: AbortSignal,
-): Promise<ResumeProjectCatalogItem[]> {
-  const data = await fetchApi<{ projects: ResumeProjectCatalogItem[] }>(
-    `/settings/rx-resumes/${encodeURIComponent(resumeId)}/projects`,
-    { signal },
-  );
-  return data.projects;
-}
-
-// Database API
-export async function clearDatabase(): Promise<{
-  message: string;
-  jobsDeleted: number;
-  runsDeleted: number;
-}> {
-  return fetchApi<{
-    message: string;
-    jobsDeleted: number;
-    runsDeleted: number;
-  }>("/database", {
-    method: "DELETE",
-  });
-}
-
-export async function deleteJobsByStatus(status: string): Promise<{
-  message: string;
-  count: number;
-}> {
-  return fetchApi<{
-    message: string;
-    count: number;
-  }>(`/jobs/status/${status}`, {
-    method: "DELETE",
-  });
-}
-
-export async function deleteJobsBelowScore(threshold: number): Promise<{
-  message: string;
-  count: number;
-  threshold: number;
-}> {
-  return fetchApi<{
-    message: string;
-    count: number;
-    threshold: number;
-  }>(`/jobs/score/${threshold}`, {
-    method: "DELETE",
-  });
-}
-
-// Visa Sponsors API
-export async function getVisaSponsorStatus(): Promise<VisaSponsorStatusResponse> {
-  return fetchApi<VisaSponsorStatusResponse>("/visa-sponsors/status");
-}
-
-export async function searchVisaSponsors(input: {
-  query: string;
-  limit?: number;
-  minScore?: number;
-  country?: string;
-}): Promise<VisaSponsorSearchResponse> {
-  if (input.query?.trim()) {
-    trackProductEvent("visa_sponsor_search", {
-      query_length_bucket: bucketQueryLength(input.query.trim()),
-      limit: input.limit,
-      min_score: input.minScore,
-      country: input.country ?? "all",
-    });
-  }
-  return fetchApi<VisaSponsorSearchResponse>("/visa-sponsors/search", {
-    method: "POST",
-    body: JSON.stringify(input),
-  });
-}
-
-export async function getVisaSponsorOrganization(
-  name: string,
-  providerId?: string,
-): Promise<VisaSponsor[]> {
-  const params = new URLSearchParams();
-  if (providerId) params.set("providerId", providerId);
-  return fetchApi<VisaSponsor[]>(
-    `/visa-sponsors/organization/${encodeURIComponent(name)}${params.size ? `?${params.toString()}` : ""}`,
-  );
-}
-
-export async function updateVisaSponsorList(): Promise<{
-  message: string;
-  status: VisaSponsorStatusResponse;
-}> {
-  return fetchApi<{
-    message: string;
-    status: VisaSponsorStatusResponse;
-  }>("/visa-sponsors/update", {
-    method: "POST",
-  });
-}
-
-// Multi-job operations (intentionally none - processing is manual)
-
-// Backup API
-export interface BackupListResponse {
-  backups: BackupInfo[];
-  nextScheduled: string | null;
-}
-
-export async function getBackups(): Promise<BackupListResponse> {
-  return fetchApi<BackupListResponse>("/backups");
-}
-
-export async function createManualBackup(): Promise<BackupInfo> {
-  return fetchApi<BackupInfo>("/backups", {
-    method: "POST",
-  });
-}
-
-export async function deleteBackup(filename: string): Promise<void> {
-  await fetchApi<void>(`/backups/${encodeURIComponent(filename)}`, {
-    method: "DELETE",
-  });
-}
+export * from "./admin";
+export * from "./auth";
+export * from "./auth-session";
+export * from "./ghostwriter";
+export * from "./jobs";
+export * from "./pipeline";
+export * from "./post-application";
+export * from "./settings-profile";
