--- orchestrator/src/server/api/routes/jobs.ts
@@ -1,2056 +1,0 @@
-import { rm } from "node:fs/promises";
-import {
-  AppError,
-  type AppErrorCode,
-  badRequest,
-  conflict,
-  notFound,
-  toAppError,
-} from "@infra/errors";
-import { fail, ok, okWithMeta } from "@infra/http";
-import { logger } from "@infra/logger";
-import { trackServerProductEvent } from "@infra/product-analytics";
-import { sanitizeWebhookPayload } from "@infra/sanitize";
-import { setupSse, startSseHeartbeat, writeSseData } from "@infra/sse";
-import { isDemoMode, sendDemoBlocked } from "@server/config/demo";
-import {
-  generateFinalPdf,
-  processJob,
-  summarizeJob,
-} from "@server/pipeline/index";
-import * as jobsRepo from "@server/repositories/jobs";
-import * as settingsRepo from "@server/repositories/settings";
-import {
-  reconcileActivationMilestonesFromHistorySafely,
-  trackCanonicalActivationEvent,
-} from "@server/services/activation-funnel";
-import {
-  deleteStageEvent,
-  getStageEvents,
-  getTasks,
-  stageEventMetadataSchema,
-  transitionStage,
-  updateStageEvent,
-} from "@server/services/applicationTracking";
-import { attachAppliedDuplicateMatches } from "@server/services/applied-duplicate-matching";
-import {
-  enqueueAutoPdfRegenerationForJob,
-  shouldEnqueueTailoringAutoPdfRegeneration,
-} from "@server/services/auto-pdf-regeneration";
-import {
-  simulateApplyJob,
-  simulateGeneratePdf,
-  simulateProcessJob,
-  simulateRescoreJob,
-  simulateSummarizeJob,
-} from "@server/services/demo-simulator";
-import { uploadJobPdf } from "@server/services/job-pdf-upload";
-import { getPdfPath, pdfExists } from "@server/services/pdf";
-import {
-  applyJobPdfFreshness,
-  applyJobsPdfFreshness,
-  type PdfFingerprintContext,
-  resolvePdfFingerprintContext,
-} from "@server/services/pdf-fingerprint";
-import { getProfile } from "@server/services/profile";
-import { scoreJobSuitability } from "@server/services/scorer";
-import { getTracerReadiness } from "@server/services/tracer-links";
-import * as visaSponsors from "@server/services/visa-sponsors/index";
-import { asyncPool } from "@server/utils/async-pool";
-import {
-  APPLICATION_OUTCOMES,
-  APPLICATION_STAGES,
-  type Job,
-  type JobAction,
-  type JobActionResponse,
-  type JobActionResult,
-  type JobActionStreamEvent,
-  type JobListItem,
-  type JobOutcome,
-  type JobStatus,
-  type JobsListResponse,
-  type JobsRevisionResponse,
-} from "@shared/types";
-import { type Request, type Response, Router } from "express";
-import { z } from "zod";
-
-export const jobsRouter = Router();
-const JOB_ACTION_CONCURRENCY = 4;
-
-const tailoredSkillsPayloadSchema = z.array(
-  z.object({
-    name: z.string(),
-    keywords: z.array(z.string()),
-  }),
-);
-
-const jobNoteSchema = z.object({
-  title: z.string().trim().min(1).max(120),
-  content: z.string().trim().min(1).max(20000),
-});
-
-async function hydrateJobPdfFreshness<T extends Job>(job: T): Promise<T> {
-  const context = await resolvePdfFingerprintContext();
-  return applyJobPdfFreshness(job, context);
-}
-
-function hydrateJobPdfFreshnessWithContext<T extends Job>(
-  job: T,
-  context: PdfFingerprintContext,
-): T {
-  return applyJobPdfFreshness(job, context);
-}
-
-function toJobListItem(
-  job: jobsRepo.JobListItemWithPdfFreshnessInput,
-): JobListItem {
-  return {
-    id: job.id,
-    source: job.source,
-    title: job.title,
-    employer: job.employer,
-    jobUrl: job.jobUrl,
-    applicationLink: job.applicationLink,
-    datePosted: job.datePosted,
-    deadline: job.deadline,
-    salary: job.salary,
-    location: job.location,
-    status: job.status,
-    outcome: job.outcome,
-    closedAt: job.closedAt,
-    suitabilityScore: job.suitabilityScore,
-    sponsorMatchScore: job.sponsorMatchScore,
-    jobType: job.jobType,
-    jobFunction: job.jobFunction,
-    pdfRegenerating: job.pdfRegenerating,
-    pdfFreshness: job.pdfFreshness,
-    salaryMinAmount: job.salaryMinAmount,
-    salaryMaxAmount: job.salaryMaxAmount,
-    salaryCurrency: job.salaryCurrency,
-    discoveredAt: job.discoveredAt,
-    readyAt: job.readyAt,
-    appliedAt: job.appliedAt,
-    updatedAt: job.updatedAt,
-  };
-}
-
-function queueTailoringAutoPdfRegenerationIfNeeded(
-  previousJob: Job,
-  nextJob: Job,
-  route: string,
-): void {
-  if (!shouldEnqueueTailoringAutoPdfRegeneration(previousJob, nextJob)) {
-    return;
-  }
-
-  queueMicrotask(() => {
-    void enqueueAutoPdfRegenerationForJob({
-      jobId: nextJob.id,
-      reason: "tailoring_updated",
-      requestedBy: "user",
-    }).catch((error) => {
-      logger.warn("Failed to queue auto PDF regeneration after job update", {
-        route,
-        jobId: nextJob.id,
-        reason: "tailoring_updated",
-        error,
-      });
-    });
-  });
-}
-
-async function notifyJobCompleteWebhook(job: Job) {
-  const overrideWebhookUrl = await settingsRepo.getSetting(
-    "jobCompleteWebhookUrl",
-  );
-  const webhookUrl = (
-    overrideWebhookUrl ||
-    process.env.JOB_COMPLETE_WEBHOOK_URL ||
-    ""
-  ).trim();
-  if (!webhookUrl) return;
-
-  try {
-    const headers: Record<string, string> = {
-      "Content-Type": "application/json",
-    };
-    const secret = process.env.WEBHOOK_SECRET;
-    if (secret) headers.Authorization = `Bearer ${secret}`;
-
-    const payload = sanitizeWebhookPayload({
-      event: "job.completed",
-      sentAt: new Date().toISOString(),
-      job: {
-        id: job.id,
-        source: job.source,
-        title: job.title,
-        employer: job.employer,
-        status: job.status,
-        suitabilityScore: job.suitabilityScore,
-        sponsorMatchScore: job.sponsorMatchScore,
-      },
-    });
-
-    const response = await fetch(webhookUrl, {
-      method: "POST",
-      headers,
-      body: JSON.stringify(payload),
-    });
-
-    if (!response.ok) {
-      logger.warn("Job complete webhook POST failed", {
-        status: response.status,
-        response: (await response.text().catch(() => "")).slice(0, 200),
-        jobId: job.id,
-      });
-    }
-  } catch (error) {
-    logger.warn("Job complete webhook POST failed", { jobId: job.id, error });
-  }
-}
-
-/**
- * PATCH /api/jobs/:id - Update a job
- */
-const updateJobSchema = z.object({
-  title: z.string().trim().min(1).max(500).optional(),
-  employer: z.string().trim().min(1).max(500).optional(),
-  jobUrl: z.string().trim().min(1).max(2000).url().optional(),
-  applicationLink: z.string().trim().max(2000).url().nullable().optional(),
-  location: z.string().trim().max(200).nullable().optional(),
-  salary: z.string().trim().max(200).nullable().optional(),
-  deadline: z.string().trim().max(100).nullable().optional(),
-  status: z
-    .enum([
-      "discovered",
-      "processing",
-      "ready",
-      "applied",
-      "in_progress",
-      "skipped",
-      "expired",
-    ])
-    .optional(),
-  outcome: z.enum(APPLICATION_OUTCOMES).nullable().optional(),
-  closedAt: z.number().int().nullable().optional(),
-  jobDescription: z.string().trim().max(40000).nullable().optional(),
-  suitabilityScore: z.number().min(0).max(100).optional(),
-  suitabilityReason: z.string().optional(),
-  tailoredSummary: z.string().optional(),
-  tailoredHeadline: z.string().optional(),
-  tailoredSkills: z
-    .string()
-    .optional()
-    .superRefine((value, ctx) => {
-      if (value === undefined || value.trim().length === 0) return;
-
-      let parsed: unknown;
-      try {
-        parsed = JSON.parse(value);
-      } catch {
-        ctx.addIssue({
-          code: z.ZodIssueCode.custom,
-          message:
-            "tailoredSkills must be a JSON array of { name, keywords } objects",
-        });
-        return;
-      }
-
-      const parseResult = tailoredSkillsPayloadSchema.safeParse(parsed);
-
-      if (!parseResult.success) {
-        ctx.addIssue({
-          code: z.ZodIssueCode.custom,
-          message:
-            "tailoredSkills must be a JSON array of { name, keywords } objects",
-        });
-      }
-    }),
-  selectedProjectIds: z.string().optional(),
-  pdfPath: z.string().optional(),
-  tracerLinksEnabled: z.boolean().optional(),
-  sponsorMatchScore: z.number().min(0).max(100).optional(),
-  sponsorMatchNames: z.string().optional(),
-});
-
-function isJobUrlConflictError(error: unknown): boolean {
-  if (!(error instanceof Error)) return false;
-  return /UNIQUE constraint failed: .*jobs\.job_url/i.test(error.message);
-}
-
-const transitionStageSchema = z.object({
-  toStage: z.enum([...APPLICATION_STAGES, "no_change"]),
-  occurredAt: z.number().int().nullable().optional(),
-  metadata: stageEventMetadataSchema.nullable().optional(),
-  outcome: z.enum(APPLICATION_OUTCOMES).nullable().optional(),
-});
-
-const updateStageEventSchema = z.object({
-  toStage: z.enum(APPLICATION_STAGES).optional(),
-  occurredAt: z.number().int().optional(),
-  metadata: stageEventMetadataSchema.nullable().optional(),
-  outcome: z.enum(APPLICATION_OUTCOMES).nullable().optional(),
-});
-
-const updateOutcomeSchema = z.object({
-  outcome: z.enum(APPLICATION_OUTCOMES).nullable(),
-  closedAt: z.number().int().nullable().optional(),
-});
-
-const jobActionRequestSchema = z.discriminatedUnion("action", [
-  z.object({
-    action: z.literal("skip"),
-    jobIds: z.array(z.string().min(1)).min(1).max(100),
-  }),
-  z.object({
-    action: z.literal("rescore"),
-    jobIds: z.array(z.string().min(1)).min(1).max(100),
-  }),
-  z.object({
-    action: z.literal("move_to_ready"),
-    jobIds: z.array(z.string().min(1)).min(1).max(100),
-    options: z
-      .object({
-        force: z.boolean().optional(),
-      })
-      .optional(),
-  }),
-]);
-
-const listJobsQuerySchema = z.object({
-  status: z.string().optional(),
-  view: z.enum(["full", "list"]).optional(),
-});
-
-const jobsRevisionQuerySchema = z.object({
-  status: z.string().optional(),
-});
-
-const uploadJobPdfSchema = z.object({
-  fileName: z.string().trim().min(1).max(255),
-  mediaType: z.string().trim().min(1).max(200).optional(),
-  dataBase64: z.string().trim().min(1),
-});
-
-const SKIPPABLE_STATUSES: ReadonlySet<JobStatus> = new Set([
-  "discovered",
-  "ready",
-]);
-const JOBS_BENCHMARK_ENABLED =
-  process.env.BENCHMARK_JOBS_TIMING === "1" ||
-  process.env.BENCHMARK_JOBS_TIMING === "true";
-
-function parseStatusFilter(statusFilter?: string): JobStatus[] | undefined {
-  const parsed = statusFilter?.split(",").filter(Boolean) as
-    | JobStatus[]
-    | undefined;
-  return parsed && parsed.length > 0 ? parsed : undefined;
-}
-
-function resolveRequestOrigin(req: Request): string | null {
-  const configuredBaseUrl = process.env.JOBOPS_PUBLIC_BASE_URL?.trim();
-  if (configuredBaseUrl) {
-    try {
-      const parsed = new URL(configuredBaseUrl);
-      if (parsed.protocol && parsed.host) {
-        return `${parsed.protocol}//${parsed.host}`;
-      }
-    } catch {
-      // Ignore invalid env and fall back to request-derived origin.
-    }
-  }
-
-  const trustProxy = Boolean(req.app?.get("trust proxy"));
-  let protocol = (req.protocol || "").trim();
-  let host = (req.header("host") || "").trim();
-
-  if (trustProxy) {
-    const forwardedProto =
-      req.header("x-forwarded-proto")?.split(",")[0]?.trim() ?? "";
-    const forwardedHost =
-      req.header("x-forwarded-host")?.split(",")[0]?.trim() ?? "";
-    if (forwardedProto) protocol = forwardedProto;
-    if (forwardedHost) host = forwardedHost;
-  }
-
-  if (!host || !protocol) return null;
-  return `${protocol}://${host}`;
-}
-
-function trackApplicationAcceptedIfNeeded(args: {
-  closedAt?: number | null;
-  nextOutcome: JobOutcome | null;
-  previousOutcome: JobOutcome | null;
-  requestOrigin?: string | null;
-  source: string;
-}): void {
-  if (
-    args.nextOutcome !== "offer_accepted" ||
-    args.previousOutcome === "offer_accepted"
-  ) {
-    return;
-  }
-
-  void trackCanonicalActivationEvent(
-    "application_accepted",
-    {
-      source: args.source,
-    },
-    {
-      occurredAt:
-        typeof args.closedAt === "number" ? args.closedAt * 1000 : Date.now(),
-      requestOrigin: args.requestOrigin ?? null,
-      urlPath: "/applications/in-progress",
-    },
-  );
-}
-
-function mapErrorForResult(error: unknown): {
-  code: string;
-  message: string;
-  details?: unknown;
-} {
-  if (error instanceof AppError) {
-    return {
-      code: error.code,
-      message: error.message,
-      ...(error.details !== undefined ? { details: error.details } : {}),
-    };
-  }
-
-  if (error instanceof Error) {
-    return {
-      code: "INTERNAL_ERROR",
-      message: error.message || "Unknown error",
-    };
-  }
-
-  return {
-    code: "INTERNAL_ERROR",
-    message: "Unknown error",
-  };
-}
-
-const STATUS_BY_APP_ERROR_CODE: Record<AppErrorCode, number> = {
-  INVALID_REQUEST: 400,
-  UNAUTHORIZED: 401,
-  FORBIDDEN: 403,
-  NOT_FOUND: 404,
-  REQUEST_TIMEOUT: 408,
-  CONFLICT: 409,
-  UNPROCESSABLE_ENTITY: 422,
-  SERVICE_UNAVAILABLE: 503,
-  UPSTREAM_ERROR: 502,
-  INTERNAL_ERROR: 500,
-};
-
-function appErrorFromPipelineFailure(
-  result: { error?: string; errorCode?: AppErrorCode },
-  fallbackMessage: string,
-): AppError {
-  const code =
-    result.errorCode ?? (result.error === "Job not found" ? "NOT_FOUND" : null);
-
-  if (!code) {
-    return badRequest(result.error ?? fallbackMessage);
-  }
-
-  return new AppError({
-    status: STATUS_BY_APP_ERROR_CODE[code],
-    code,
-    message: result.error ?? fallbackMessage,
-  });
-}
-
-type JobActionExecutionOptions = {
-  getProfileForRescore?: () => Promise<Record<string, unknown>>;
-  forceMoveToReady?: boolean;
-  requestOrigin?: string | null;
-  analyticsOrigin?:
-    | "move_to_ready"
-    | "generate_pdf"
-    | "pipeline"
-    | "manual_job_create";
-};
-
-function createSharedRescoreProfileLoader(): () => Promise<
-  Record<string, unknown>
-> {
-  let profilePromise: Promise<Record<string, unknown>> | null = null;
-
-  return async () => {
-    if (!profilePromise) {
-      profilePromise = (async () => {
-        const rawProfile = await getProfile();
-        if (
-          !rawProfile ||
-          typeof rawProfile !== "object" ||
-          Array.isArray(rawProfile)
-        ) {
-          throw badRequest("Invalid resume profile format");
-        }
-        return rawProfile as Record<string, unknown>;
-      })();
-    }
-    return profilePromise;
-  };
-}
-
-async function executeJobActionForJob(
-  action: JobAction,
-  jobId: string,
-  options?: JobActionExecutionOptions,
-): Promise<JobActionResult> {
-  try {
-    const job = await jobsRepo.getJobById(jobId);
-    if (!job) {
-      throw new AppError({
-        status: 404,
-        code: "NOT_FOUND",
-        message: "Job not found",
-      });
-    }
-
-    if (action === "skip") {
-      if (!SKIPPABLE_STATUSES.has(job.status)) {
-        throw badRequest(`Job is not skippable from status "${job.status}"`, {
-          jobId,
-          status: job.status,
-          allowedStatuses: ["discovered", "ready"],
-        });
-      }
-
-      const updated = await jobsRepo.updateJob(jobId, { status: "skipped" });
-      if (!updated) {
-        throw new AppError({
-          status: 404,
-          code: "NOT_FOUND",
-          message: "Job not found",
-        });
-      }
-
-      return { jobId, ok: true, job: updated };
-    }
-
-    if (action === "move_to_ready") {
-      if (job.status !== "discovered") {
-        throw badRequest(
-          `Job is not movable to Ready from status "${job.status}"`,
-          {
-            jobId,
-            status: job.status,
-            requiredStatus: "discovered",
-          },
-        );
-      }
-
-      if (isDemoMode()) {
-        const simulated = await simulateProcessJob(jobId, {
-          force: options?.forceMoveToReady ?? false,
-        });
-        if (!simulated.success) {
-          throw new AppError({
-            status: 500,
-            code: "INTERNAL_ERROR",
-            message: simulated.error || "Failed to process job",
-          });
-        }
-      } else {
-        const processed = await processJob(jobId, {
-          force: options?.forceMoveToReady ?? false,
-          requestOrigin: options?.requestOrigin ?? null,
-          analyticsOrigin: options?.analyticsOrigin ?? "move_to_ready",
-        });
-        if (!processed.success) {
-          throw new AppError({
-            status: 500,
-            code: "INTERNAL_ERROR",
-            message: processed.error || "Failed to process job",
-          });
-        }
-      }
-
-      const updated = await jobsRepo.getJobById(jobId);
-      if (!updated) {
-        throw new AppError({
-          status: 404,
-          code: "NOT_FOUND",
-          message: "Job not found after processing",
-        });
-      }
-
-      return { jobId, ok: true, job: updated };
-    }
-
-    if (job.status === "processing") {
-      throw badRequest(`Job is not rescorable from status "${job.status}"`, {
-        jobId,
-        status: job.status,
-        disallowedStatus: "processing",
-      });
-    }
-
-    if (isDemoMode()) {
-      const simulated = await simulateRescoreJob(job.id);
-      return { jobId, ok: true, job: simulated };
-    }
-
-    const profile = options?.getProfileForRescore
-      ? await options.getProfileForRescore()
-      : await (async () => {
-          const rawProfile = await getProfile();
-          if (
-            !rawProfile ||
-            typeof rawProfile !== "object" ||
-            Array.isArray(rawProfile)
-          ) {
-            throw badRequest("Invalid resume profile format");
-          }
-          return rawProfile as Record<string, unknown>;
-        })();
-
-    const { score, reason } = await scoreJobSuitability(job, profile);
-
-    const updated = await jobsRepo.updateJob(job.id, {
-      suitabilityScore: score,
-      suitabilityReason: reason,
-    });
-    if (!updated) {
-      throw new AppError({
-        status: 404,
-        code: "NOT_FOUND",
-        message: "Job not found",
-      });
-    }
-
-    return { jobId, ok: true, job: updated };
-  } catch (error) {
-    const mapped = mapErrorForResult(error);
-    return {
-      jobId,
-      ok: false,
-      error: {
-        code: mapped.code,
-        message: mapped.message,
-      },
-    };
-  }
-}
-
-function mapJobActionFailure(
-  failure: Extract<JobActionResult, { ok: false }>,
-): AppError {
-  const code = (
-    failure.error.code in STATUS_BY_APP_ERROR_CODE
-      ? failure.error.code
-      : "INTERNAL_ERROR"
-  ) as AppErrorCode;
-
-  return new AppError({
-    status: STATUS_BY_APP_ERROR_CODE[code],
-    code,
-    message: failure.error.message,
-  });
-}
-
-/**
- * GET /api/jobs - List all jobs
- * Query params: status (comma-separated list of statuses to filter)
- */
-jobsRouter.get("/", async (req: Request, res: Response) => {
-  try {
-    const benchmarkStart = performance.now();
-    let queryParseMs = 0;
-    let primaryQueryMs = 0;
-    const duplicateCandidatesQueryMs = 0;
-    const duplicateMatchCpuMs = 0;
-    let statsAggregateMs = 0;
-    let revisionAggregateMs = 0;
-
-    const queryParseStart = performance.now();
-    const parsedQuery = listJobsQuerySchema.safeParse(req.query);
-    queryParseMs = performance.now() - queryParseStart;
-    if (!parsedQuery.success) {
-      return fail(
-        res,
-        badRequest(
-          "Invalid jobs list query parameters",
-          parsedQuery.error.flatten(),
-        ),
-      );
-    }
-
-    const statusFilter = parsedQuery.data.status;
-    const statuses = parseStatusFilter(statusFilter);
-    const view = parsedQuery.data.view ?? "list";
-
-    const primaryQueryStart = performance.now();
-    const pdfFingerprintContext = await resolvePdfFingerprintContext();
-    const jobs: Array<Job | JobListItem> =
-      view === "list"
-        ? applyJobsPdfFreshness(
-            await jobsRepo.getJobListItems(statuses),
-            pdfFingerprintContext,
-          ).map(toJobListItem)
-        : applyJobsPdfFreshness(
-            await jobsRepo.getAllJobs(statuses),
-            pdfFingerprintContext,
-          );
-    primaryQueryMs = performance.now() - primaryQueryStart;
-    const candidateCount = 0;
-    const duplicateMatchingEnabled = false;
-    const statsAggregateStart = performance.now();
-    const stats = await jobsRepo.getJobStats();
-    statsAggregateMs = performance.now() - statsAggregateStart;
-    const revisionAggregateStart = performance.now();
-    const revision = await jobsRepo.getJobsRevision(statuses);
-    revisionAggregateMs = performance.now() - revisionAggregateStart;
-
-    const response: JobsListResponse<Job | JobListItem> = {
-      jobs,
-      total: jobs.length,
-      byStatus: stats,
-      revision: revision.revision,
-    };
-    const internalRouteMs =
-      queryParseMs +
-      primaryQueryMs +
-      duplicateCandidatesQueryMs +
-      duplicateMatchCpuMs +
-      statsAggregateMs +
-      revisionAggregateMs;
-    const totalMs = performance.now() - benchmarkStart;
-
-    if (JOBS_BENCHMARK_ENABLED) {
-      logger.info("Jobs list benchmark", {
-        route: "GET /api/jobs",
-        view,
-        statusFilter: statusFilter ?? null,
-        returnedCount: jobs.length,
-        duplicateMatchingEnabled,
-        candidateCount,
-        totalMs,
-        queryParseMs,
-        primaryQueryMs,
-        duplicateCandidatesQueryMs,
-        duplicateMatchCpuMs,
-        statsAggregateMs,
-        revisionAggregateMs,
-        internalRouteMs,
-      });
-    }
-
-    logger.info("Jobs list fetched", {
-      route: "GET /api/jobs",
-      view,
-      statusFilter: statusFilter ?? null,
-      revision: revision.revision,
-      returnedCount: jobs.length,
-    });
-
-    ok(res, response);
-  } catch (error) {
-    const err =
-      error instanceof AppError
-        ? error
-        : new AppError({
-            status: 500,
-            code: "INTERNAL_ERROR",
-            message: error instanceof Error ? error.message : "Unknown error",
-          });
-    fail(res, err);
-  }
-});
-
-/**
- * GET /api/jobs/revision - Get jobs list revision for lightweight change detection
- * Query params: status (comma-separated list of statuses to filter)
- */
-jobsRouter.get("/revision", async (req: Request, res: Response) => {
-  try {
-    const parsedQuery = jobsRevisionQuerySchema.safeParse(req.query);
-    if (!parsedQuery.success) {
-      return fail(
-        res,
-        badRequest(
-          "Invalid jobs revision query parameters",
-          parsedQuery.error.flatten(),
-        ),
-      );
-    }
-
-    const statuses = parseStatusFilter(parsedQuery.data.status);
-    const revision = await jobsRepo.getJobsRevision(statuses);
-
-    const response: JobsRevisionResponse = {
-      revision: revision.revision,
-      latestUpdatedAt: revision.latestUpdatedAt,
-      total: revision.total,
-      statusFilter: revision.statusFilter,
-    };
-
-    logger.info("Jobs revision fetched", {
-      route: "GET /api/jobs/revision",
-      statusFilter: revision.statusFilter,
-      revision: revision.revision,
-      total: revision.total,
-    });
-
-    ok(res, response);
-  } catch (error) {
-    const err =
-      error instanceof AppError
-        ? error
-        : new AppError({
-            status: 500,
-            code: "INTERNAL_ERROR",
-            message: error instanceof Error ? error.message : "Unknown error",
-          });
-    fail(res, err);
-  }
-});
-
-/**
- * POST /api/jobs/actions - Run a job action across selected jobs
- */
-jobsRouter.post("/actions", async (req: Request, res: Response) => {
-  try {
-    const parsed = jobActionRequestSchema.parse(req.body);
-    const dedupedJobIds = Array.from(new Set(parsed.jobIds));
-    const requestOrigin = resolveRequestOrigin(req);
-    const executionOptions: JobActionExecutionOptions = {
-      ...(parsed.action === "rescore" && !isDemoMode()
-        ? { getProfileForRescore: createSharedRescoreProfileLoader() }
-        : {}),
-      ...(parsed.action === "move_to_ready" &&
-      parsed.options?.force !== undefined
-        ? { forceMoveToReady: parsed.options.force }
-        : {}),
-      ...(parsed.action === "move_to_ready" ? { requestOrigin } : {}),
-    };
-
-    const rawResults = await asyncPool({
-      items: dedupedJobIds,
-      concurrency: JOB_ACTION_CONCURRENCY,
-      task: async (jobId) =>
-        executeJobActionForJob(parsed.action, jobId, executionOptions),
-    });
-    const pdfFingerprintContext = await resolvePdfFingerprintContext();
-    const results = await Promise.all(
-      rawResults.map(async (result) =>
-        result.ok
-          ? {
-              ...result,
-              job: hydrateJobPdfFreshnessWithContext(
-                result.job,
-                pdfFingerprintContext,
-              ),
-            }
-          : result,
-      ),
-    );
-
-    const succeeded = results.filter((result) => result.ok).length;
-    const failed = results.length - succeeded;
-    const payload: JobActionResponse = {
-      action: parsed.action,
-      requested: dedupedJobIds.length,
-      succeeded,
-      failed,
-      results,
-    };
-
-    logger.info("Job action completed", {
-      route: "POST /api/jobs/actions",
-      action: parsed.action,
-      requested: dedupedJobIds.length,
-      succeeded,
-      failed,
-      concurrency: JOB_ACTION_CONCURRENCY,
-    });
-
-    ok(res, payload);
-  } catch (error) {
-    const err =
-      error instanceof z.ZodError
-        ? badRequest("Invalid job action request", error.flatten())
-        : error instanceof AppError
-          ? error
-          : new AppError({
-              status: 500,
-              code: "INTERNAL_ERROR",
-              message: error instanceof Error ? error.message : "Unknown error",
-            });
-
-    logger.error("Job action failed", {
-      route: "POST /api/jobs/actions",
-      status: err.status,
-      code: err.code,
-      details: err.details,
-    });
-
-    fail(res, err);
-  }
-});
-
-/**
- * POST /api/jobs/actions/stream - Run a job action and stream per-job progress via SSE
- */
-jobsRouter.post("/actions/stream", async (req: Request, res: Response) => {
-  const parsed = jobActionRequestSchema.safeParse(req.body);
-  if (!parsed.success) {
-    return fail(
-      res,
-      badRequest("Invalid job action request", parsed.error.flatten()),
-    );
-  }
-
-  const dedupedJobIds = Array.from(new Set(parsed.data.jobIds));
-  const requestOrigin = resolveRequestOrigin(req);
-  const requestId = String(res.getHeader("x-request-id") || "unknown");
-  const action = parsed.data.action;
-  const executionOptions: JobActionExecutionOptions = {
-    ...(action === "rescore" && !isDemoMode()
-      ? { getProfileForRescore: createSharedRescoreProfileLoader() }
-      : {}),
-    ...(action === "move_to_ready" && parsed.data.options?.force !== undefined
-      ? { forceMoveToReady: parsed.data.options.force }
-      : {}),
-    ...(action === "move_to_ready" ? { requestOrigin } : {}),
-  };
-  const requested = dedupedJobIds.length;
-  const results: JobActionResult[] = [];
-  let succeeded = 0;
-  let failed = 0;
-
-  setupSse(res, {
-    cacheControl: "no-cache, no-transform",
-    disableBuffering: true,
-    flushHeaders: true,
-  });
-  const stopHeartbeat = startSseHeartbeat(res);
-
-  let clientDisconnected = false;
-  res.on("close", () => {
-    clientDisconnected = true;
-    stopHeartbeat();
-  });
-
-  const isResponseWritable = () =>
-    !clientDisconnected && !res.writableEnded && !res.destroyed;
-
-  const sendEvent = (event: JobActionStreamEvent) => {
-    if (!isResponseWritable()) return false;
-    writeSseData(res, event);
-    return true;
-  };
-
-  try {
-    const pdfFingerprintContext = await resolvePdfFingerprintContext();
-
-    if (
-      !sendEvent({
-        type: "started",
-        action,
-        requested,
-        completed: 0,
-        succeeded: 0,
-        failed: 0,
-        requestId,
-      })
-    ) {
-      logger.info("Client disconnected before action stream started", {
-        route: "POST /api/jobs/actions/stream",
-        action,
-        requested,
-        succeeded,
-        failed,
-        requestId,
-      });
-      return;
-    }
-
-    await asyncPool({
-      items: dedupedJobIds,
-      concurrency: JOB_ACTION_CONCURRENCY,
-      shouldStop: () => !isResponseWritable(),
-      task: async (jobId) => {
-        if (!isResponseWritable()) return;
-
-        const rawResult = await executeJobActionForJob(
-          action,
-          jobId,
-          executionOptions,
-        );
-        const result = rawResult.ok
-          ? {
-              ...rawResult,
-              job: hydrateJobPdfFreshnessWithContext(
-                rawResult.job,
-                pdfFingerprintContext,
-              ),
-            }
-          : rawResult;
-        results.push(result);
-        if (result.ok) succeeded += 1;
-        else failed += 1;
-
-        if (
-          !sendEvent({
-            type: "progress",
-            action,
-            requested,
-            completed: results.length,
-            succeeded,
-            failed,
-            result,
-            requestId,
-          })
-        ) {
-          logger.info(
-            "Client disconnected while writing action stream progress",
-            {
-              route: "POST /api/jobs/actions/stream",
-              action,
-              requested,
-              succeeded,
-              failed,
-              requestId,
-            },
-          );
-        }
-      },
-    });
-
-    sendEvent({
-      type: "completed",
-      action,
-      requested,
-      completed: results.length,
-      succeeded,
-      failed,
-      results,
-      requestId,
-    });
-
-    logger.info("Job action stream completed", {
-      route: "POST /api/jobs/actions/stream",
-      action,
-      requested,
-      succeeded,
-      failed,
-      concurrency: JOB_ACTION_CONCURRENCY,
-      requestId,
-    });
-  } catch (error) {
-    const err =
-      error instanceof AppError
-        ? error
-        : new AppError({
-            status: 500,
-            code: "INTERNAL_ERROR",
-            message: error instanceof Error ? error.message : "Unknown error",
-          });
-
-    logger.error("Job action stream failed", {
-      route: "POST /api/jobs/actions/stream",
-      action,
-      requested,
-      succeeded,
-      failed,
-      status: err.status,
-      code: err.code,
-      requestId,
-    });
-
-    if (
-      !sendEvent({
-        type: "error",
-        code: err.code,
-        message: err.message,
-        requestId,
-      })
-    ) {
-      logger.info("Skipping stream error event because client disconnected", {
-        route: "POST /api/jobs/actions/stream",
-        action,
-        requested,
-        succeeded,
-        failed,
-        requestId,
-      });
-    }
-  } finally {
-    stopHeartbeat();
-    if (!res.writableEnded && !res.destroyed) {
-      res.end();
-    }
-  }
-});
-
-jobsRouter.post("/:id/process", async (req: Request, res: Response) => {
-  const forceRaw = req.query.force as string | undefined;
-  const force = forceRaw === "1" || forceRaw === "true";
-  const result = await executeJobActionForJob("move_to_ready", req.params.id, {
-    forceMoveToReady: force,
-    requestOrigin: resolveRequestOrigin(req),
-  });
-  if (!result.ok) return fail(res, mapJobActionFailure(result));
-  ok(res, await hydrateJobPdfFreshness(result.job));
-});
-
-jobsRouter.post("/:id/skip", async (req: Request, res: Response) => {
-  const result = await executeJobActionForJob("skip", req.params.id);
-  if (!result.ok) return fail(res, mapJobActionFailure(result));
-  ok(res, await hydrateJobPdfFreshness(result.job));
-});
-
-jobsRouter.post("/:id/rescore", async (req: Request, res: Response) => {
-  const result = await executeJobActionForJob("rescore", req.params.id, {
-    ...(isDemoMode()
-      ? {}
-      : { getProfileForRescore: createSharedRescoreProfileLoader() }),
-  });
-  if (!result.ok) return fail(res, mapJobActionFailure(result));
-  ok(res, await hydrateJobPdfFreshness(result.job));
-});
-
-/**
- * GET /api/jobs/:id - Get a single job
- */
-jobsRouter.get("/:id", async (req: Request, res: Response) => {
-  try {
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      return fail(res, notFound("Job not found"));
-    }
-    const [jobWithAppliedDuplicateMatch] = attachAppliedDuplicateMatches(
-      [job],
-      await jobsRepo.getAppliedDuplicateMatchCandidates(),
-    );
-    ok(res, await hydrateJobPdfFreshness(jobWithAppliedDuplicateMatch));
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * GET /api/jobs/:id/events - Get stage event timeline
- */
-jobsRouter.get("/:id/events", async (req: Request, res: Response) => {
-  try {
-    const events = await getStageEvents(req.params.id);
-    ok(res, events);
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * GET /api/jobs/:id/notes - Get notes for a job
- */
-jobsRouter.get("/:id/notes", async (req: Request, res: Response) => {
-  const requestId = String(res.getHeader("x-request-id") || "unknown");
-
-  try {
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      const err = notFound("Job not found");
-      logger.warn("Job notes fetch failed", {
-        route: "GET /api/jobs/:id/notes",
-        jobId: req.params.id,
-        requestId,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    const notes = await jobsRepo.listJobNotes(job.id);
-
-    logger.info("Job notes fetched", {
-      route: "GET /api/jobs/:id/notes",
-      jobId: job.id,
-      requestId,
-      returnedCount: notes.length,
-    });
-
-    ok(res, notes);
-  } catch (error) {
-    const err = toAppError(error);
-    logger.error("Job notes fetch failed", {
-      route: "GET /api/jobs/:id/notes",
-      jobId: req.params.id,
-      requestId,
-      status: err.status,
-      code: err.code,
-      details: err.details,
-      errorMessage: error instanceof Error ? error.message : undefined,
-    });
-    fail(res, err);
-  }
-});
-
-/**
- * POST /api/jobs/:id/notes - Create a note for a job
- */
-jobsRouter.post("/:id/notes", async (req: Request, res: Response) => {
-  const requestId = String(res.getHeader("x-request-id") || "unknown");
-
-  try {
-    const input = jobNoteSchema.safeParse(req.body);
-    if (!input.success) {
-      return fail(
-        res,
-        badRequest("Invalid job note request", input.error.flatten()),
-      );
-    }
-
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      const err = notFound("Job not found");
-      logger.warn("Job note create failed", {
-        route: "POST /api/jobs/:id/notes",
-        jobId: req.params.id,
-        requestId,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    const note = await jobsRepo.createJobNote({
-      jobId: job.id,
-      ...input.data,
-    });
-
-    logger.info("Job note created", {
-      route: "POST /api/jobs/:id/notes",
-      jobId: job.id,
-      noteId: note.id,
-      requestId,
-    });
-
-    ok(res, note, 201);
-  } catch (error) {
-    const err = toAppError(error);
-    logger.error("Job note create failed", {
-      route: "POST /api/jobs/:id/notes",
-      jobId: req.params.id,
-      requestId,
-      status: err.status,
-      code: err.code,
-      details: err.details,
-      errorMessage: error instanceof Error ? error.message : undefined,
-    });
-    fail(res, err);
-  }
-});
-
-/**
- * PATCH /api/jobs/:id/notes/:noteId - Update a job note
- */
-jobsRouter.patch("/:id/notes/:noteId", async (req: Request, res: Response) => {
-  const requestId = String(res.getHeader("x-request-id") || "unknown");
-
-  try {
-    const input = jobNoteSchema.safeParse(req.body);
-    if (!input.success) {
-      return fail(
-        res,
-        badRequest("Invalid job note request", input.error.flatten()),
-      );
-    }
-
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      const err = notFound("Job not found");
-      logger.warn("Job note update failed", {
-        route: "PATCH /api/jobs/:id/notes/:noteId",
-        jobId: req.params.id,
-        noteId: req.params.noteId,
-        requestId,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    const note = await jobsRepo.updateJobNote({
-      jobId: job.id,
-      noteId: req.params.noteId,
-      ...input.data,
-    });
-    if (!note) {
-      const err = notFound("Job note not found");
-      logger.warn("Job note update failed", {
-        route: "PATCH /api/jobs/:id/notes/:noteId",
-        jobId: job.id,
-        noteId: req.params.noteId,
-        requestId,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    logger.info("Job note updated", {
-      route: "PATCH /api/jobs/:id/notes/:noteId",
-      jobId: job.id,
-      noteId: note.id,
-      requestId,
-    });
-
-    ok(res, note);
-  } catch (error) {
-    const err = toAppError(error);
-    logger.error("Job note update failed", {
-      route: "PATCH /api/jobs/:id/notes/:noteId",
-      jobId: req.params.id,
-      noteId: req.params.noteId,
-      requestId,
-      status: err.status,
-      code: err.code,
-      details: err.details,
-      errorMessage: error instanceof Error ? error.message : undefined,
-    });
-    fail(res, err);
-  }
-});
-
-/**
- * DELETE /api/jobs/:id/notes/:noteId - Delete a job note
- */
-jobsRouter.delete("/:id/notes/:noteId", async (req: Request, res: Response) => {
-  const requestId = String(res.getHeader("x-request-id") || "unknown");
-
-  try {
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      const err = notFound("Job not found");
-      logger.warn("Job note delete failed", {
-        route: "DELETE /api/jobs/:id/notes/:noteId",
-        jobId: req.params.id,
-        noteId: req.params.noteId,
-        requestId,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    const deletedCount = await jobsRepo.deleteJobNote({
-      jobId: job.id,
-      noteId: req.params.noteId,
-    });
-    if (deletedCount === 0) {
-      const err = notFound("Job note not found");
-      logger.warn("Job note delete failed", {
-        route: "DELETE /api/jobs/:id/notes/:noteId",
-        jobId: job.id,
-        noteId: req.params.noteId,
-        requestId,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    logger.info("Job note deleted", {
-      route: "DELETE /api/jobs/:id/notes/:noteId",
-      jobId: job.id,
-      noteId: req.params.noteId,
-      requestId,
-    });
-
-    ok(res, null);
-  } catch (error) {
-    const err = toAppError(error);
-    logger.error("Job note delete failed", {
-      route: "DELETE /api/jobs/:id/notes/:noteId",
-      jobId: req.params.id,
-      noteId: req.params.noteId,
-      requestId,
-      status: err.status,
-      code: err.code,
-      details: err.details,
-      errorMessage: error instanceof Error ? error.message : undefined,
-    });
-    fail(res, err);
-  }
-});
-
-/**
- * GET /api/jobs/:id/tasks - Get tasks for an application
- */
-jobsRouter.get("/:id/tasks", async (req: Request, res: Response) => {
-  try {
-    const includeCompleted =
-      req.query.includeCompleted === "1" ||
-      req.query.includeCompleted === "true";
-    const tasks = await getTasks(req.params.id, includeCompleted);
-    ok(res, tasks);
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * POST /api/jobs/:id/stages - Transition stage
- */
-jobsRouter.post("/:id/stages", async (req: Request, res: Response) => {
-  try {
-    const input = transitionStageSchema.parse(req.body);
-    const event = transitionStage(
-      req.params.id,
-      input.toStage,
-      input.occurredAt ?? undefined,
-      input.metadata ?? null,
-      input.outcome ?? null,
-    );
-    ok(res, event);
-    queueMicrotask(() => {
-      void reconcileActivationMilestonesFromHistorySafely({
-        route: "POST /api/jobs/:id/stages",
-        jobId: req.params.id,
-      });
-    });
-  } catch (error) {
-    if (error instanceof z.ZodError) {
-      return fail(res, badRequest(error.message, error.flatten()));
-    }
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * PATCH /api/jobs/:id/events/:eventId - Update an event
- */
-jobsRouter.patch(
-  "/:id/events/:eventId",
-  async (req: Request, res: Response) => {
-    try {
-      const input = updateStageEventSchema.parse(req.body);
-      updateStageEvent(req.params.eventId, input);
-      ok(res, null);
-      queueMicrotask(() => {
-        void reconcileActivationMilestonesFromHistorySafely({
-          route: "PATCH /api/jobs/:id/events/:eventId",
-          jobId: req.params.id,
-          eventId: req.params.eventId,
-        });
-      });
-    } catch (error) {
-      if (error instanceof z.ZodError) {
-        return fail(res, badRequest(error.message, error.flatten()));
-      }
-      fail(res, toAppError(error));
-    }
-  },
-);
-
-/**
- * DELETE /api/jobs/:id/events/:eventId - Delete an event
- */
-jobsRouter.delete(
-  "/:id/events/:eventId",
-  async (req: Request, res: Response) => {
-    try {
-      deleteStageEvent(req.params.eventId);
-      ok(res, null);
-      queueMicrotask(() => {
-        void reconcileActivationMilestonesFromHistorySafely({
-          route: "DELETE /api/jobs/:id/events/:eventId",
-          jobId: req.params.id,
-          eventId: req.params.eventId,
-        });
-      });
-    } catch (error) {
-      fail(res, toAppError(error));
-    }
-  },
-);
-
-/**
- * PATCH /api/jobs/:id/outcome - Close out application
- */
-jobsRouter.patch("/:id/outcome", async (req: Request, res: Response) => {
-  try {
-    const input = updateOutcomeSchema.parse(req.body);
-    const currentJob = await jobsRepo.getJobById(req.params.id);
-    if (!currentJob) {
-      return fail(res, notFound("Job not found"));
-    }
-    const closedAt = input.outcome
-      ? (input.closedAt ?? Math.floor(Date.now() / 1000))
-      : null;
-    const job = await jobsRepo.updateJob(req.params.id, {
-      outcome: input.outcome,
-      closedAt,
-    });
-
-    if (!job) {
-      return fail(res, notFound("Job not found"));
-    }
-
-    trackApplicationAcceptedIfNeeded({
-      closedAt: job.closedAt,
-      nextOutcome: job.outcome,
-      previousOutcome: currentJob.outcome,
-      requestOrigin: resolveRequestOrigin(req),
-      source: "jobs_outcome_route",
-    });
-
-    ok(res, await hydrateJobPdfFreshness(job));
-    queueMicrotask(() => {
-      void reconcileActivationMilestonesFromHistorySafely({
-        route: "PATCH /api/jobs/:id/outcome",
-        jobId: req.params.id,
-      });
-    });
-  } catch (error) {
-    if (error instanceof z.ZodError) {
-      return fail(res, badRequest(error.message, error.flatten()));
-    }
-    fail(res, toAppError(error));
-  }
-});
-
-jobsRouter.patch("/:id", async (req: Request, res: Response) => {
-  try {
-    const input = updateJobSchema.parse(req.body);
-    const currentJob = await jobsRepo.getJobById(req.params.id);
-
-    if (!currentJob) {
-      const err = new AppError({
-        status: 404,
-        code: "NOT_FOUND",
-        message: "Job not found",
-      });
-      logger.warn("Job update failed", {
-        route: "PATCH /api/jobs/:id",
-        jobId: req.params.id,
-        status: err.status,
-        code: err.code,
-      });
-      fail(res, err);
-      return;
-    }
-
-    const isTurningTracerLinksOn =
-      input.tracerLinksEnabled === true && !currentJob.tracerLinksEnabled;
-
-    if (isTurningTracerLinksOn) {
-      const readiness = await getTracerReadiness({
-        requestOrigin: resolveRequestOrigin(req),
-        force: true,
-      });
-
-      if (!readiness.canEnable) {
-        throw new AppError({
-          status: 409,
-          code: "CONFLICT",
-          message:
-            readiness.reason ??
-            "Tracer links are unavailable right now. Verify Tracer Links in Settings.",
-          details: {
-            tracerReadiness: {
-              status: readiness.status,
-              checkedAt: readiness.checkedAt,
-              publicBaseUrl: readiness.publicBaseUrl,
-            },
-          },
-        });
-      }
-    }
-
-    const job = await jobsRepo.updateJob(req.params.id, input);
-
-    if (!job) {
-      const err = new AppError({
-        status: 404,
-        code: "NOT_FOUND",
-        message: "Job not found",
-      });
-      logger.warn("Job update failed", {
-        route: "PATCH /api/jobs/:id",
-        jobId: req.params.id,
-        status: err.status,
-        code: err.code,
-      });
-      return fail(res, err);
-    }
-
-    logger.info("Job updated", {
-      route: "PATCH /api/jobs/:id",
-      jobId: req.params.id,
-      updatedFields: Object.keys(input),
-    });
-
-    trackApplicationAcceptedIfNeeded({
-      closedAt: job.closedAt,
-      nextOutcome: job.outcome,
-      previousOutcome: currentJob.outcome,
-      requestOrigin: resolveRequestOrigin(req),
-      source: "jobs_patch_route",
-    });
-    ok(res, await hydrateJobPdfFreshness(job));
-    queueTailoringAutoPdfRegenerationIfNeeded(
-      currentJob,
-      job,
-      "PATCH /api/jobs/:id",
-    );
-    if (
-      Object.hasOwn(input, "appliedAt") ||
-      Object.hasOwn(input, "closedAt") ||
-      Object.hasOwn(input, "outcome")
-    ) {
-      queueMicrotask(() => {
-        void reconcileActivationMilestonesFromHistorySafely({
-          route: "PATCH /api/jobs/:id",
-          jobId: req.params.id,
-          updatedFields: Object.keys(input),
-        });
-      });
-    }
-  } catch (error) {
-    const err =
-      error instanceof z.ZodError
-        ? badRequest(
-            error.issues[0]?.message ?? "Invalid job update request",
-            error.flatten(),
-          )
-        : isJobUrlConflictError(error)
-          ? conflict("Another job already uses that job URL")
-          : error instanceof AppError
-            ? error
-            : new AppError({
-                status: 500,
-                code: "INTERNAL_ERROR",
-                message:
-                  error instanceof Error ? error.message : "Unknown error",
-              });
-
-    logger.error("Job update failed", {
-      route: "PATCH /api/jobs/:id",
-      jobId: req.params.id,
-      status: err.status,
-      code: err.code,
-      details: err.details,
-    });
-
-    fail(res, err);
-  }
-});
-
-jobsRouter.post("/:id/pdf", async (req: Request, res: Response) => {
-  let uploadedPath: string | null = null;
-
-  try {
-    const input = uploadJobPdfSchema.parse(req.body);
-    const currentJob = await jobsRepo.getJobById(req.params.id);
-
-    if (!currentJob) {
-      const err = new AppError({
-        status: 404,
-        code: "NOT_FOUND",
-        message: "Job not found",
-      });
-      logger.warn("Job PDF upload failed", {
-        route: "POST /api/jobs/:id/pdf",
-        jobId: req.params.id,
-        status: err.status,
-        code: err.code,
-      });
-      fail(res, err);
-      return;
-    }
-
-    const uploaded = await uploadJobPdf({
-      jobId: req.params.id,
-      fileName: input.fileName,
-      mediaType: input.mediaType,
-      dataBase64: input.dataBase64,
-    });
-    uploadedPath = uploaded.outputPath;
-
-    const job = await jobsRepo.updateJob(req.params.id, {
-      pdfPath: uploaded.outputPath,
-      pdfSource: "uploaded",
-      pdfRegenerating: false,
-      pdfFingerprint: null,
-      pdfGeneratedAt: new Date().toISOString(),
-    });
-
-    if (!job) {
-      await rm(uploaded.outputPath, { force: true }).catch((cleanupError) => {
-        logger.warn("Failed to clean up uploaded PDF after missing job", {
-          route: "POST /api/jobs/:id/pdf",
-          jobId: req.params.id,
-          cleanupError,
-        });
-      });
-
-      const err = new AppError({
-        status: 404,
-        code: "NOT_FOUND",
-        message: "Job not found",
-      });
-      logger.warn("Job PDF upload failed", {
-        route: "POST /api/jobs/:id/pdf",
-        jobId: req.params.id,
-        status: err.status,
-        code: err.code,
-      });
-      fail(res, err);
-      return;
-    }
-
-    logger.info("Job PDF uploaded", {
-      route: "POST /api/jobs/:id/pdf",
-      jobId: req.params.id,
-      fileName: input.fileName,
-      byteLength: uploaded.byteLength,
-    });
-
-    ok(res, await hydrateJobPdfFreshness(job), 201);
-  } catch (error) {
-    const err =
-      error instanceof z.ZodError
-        ? badRequest(
-            error.issues[0]?.message ?? "Invalid job PDF upload request",
-            error.flatten(),
-          )
-        : error instanceof AppError
-          ? error
-          : new AppError({
-              status: 500,
-              code: "INTERNAL_ERROR",
-              message: error instanceof Error ? error.message : "Unknown error",
-            });
-
-    if (uploadedPath) {
-      await rm(uploadedPath, { force: true }).catch((cleanupError) => {
-        logger.warn("Failed to clean up uploaded PDF after route error", {
-          route: "POST /api/jobs/:id/pdf",
-          jobId: req.params.id,
-          cleanupError,
-        });
-      });
-    }
-
-    logger.error("Job PDF upload failed", {
-      route: "POST /api/jobs/:id/pdf",
-      jobId: req.params.id,
-      status: err.status,
-      code: err.code,
-      details: err.details,
-      uploadedPath,
-    });
-
-    fail(res, err);
-  }
-});
-
-jobsRouter.get("/:id/pdf", async (req: Request, res: Response) => {
-  const currentJob = await jobsRepo.getJobById(req.params.id);
-  if (!currentJob || !(await pdfExists(req.params.id))) {
-    fail(res, notFound("PDF not found"));
-    return;
-  }
-
-  const pdfPath = getPdfPath(req.params.id);
-  res.setHeader("Cache-Control", "no-store");
-  res.sendFile(pdfPath, (error) => {
-    if (error) {
-      fail(res, notFound("PDF not found"));
-    }
-  });
-});
-
-/**
- * POST /api/jobs/:id/summarize - Generate AI summary and suggest projects
- */
-jobsRouter.post("/:id/summarize", async (req: Request, res: Response) => {
-  try {
-    const forceRaw = req.query.force as string | undefined;
-    const force = forceRaw === "1" || forceRaw === "true";
-
-    if (isDemoMode()) {
-      const result = await simulateSummarizeJob(req.params.id, { force });
-      if (!result.success) {
-        return fail(
-          res,
-          badRequest(result.error ?? "Failed to summarize the job"),
-        );
-      }
-      const job = await jobsRepo.getJobById(req.params.id);
-      if (!job) {
-        return fail(res, notFound("Job not found"));
-      }
-      return okWithMeta(res, await hydrateJobPdfFreshness(job), {
-        simulated: true,
-      });
-    }
-
-    const previousJob = await jobsRepo.getJobById(req.params.id);
-    const result = await summarizeJob(req.params.id, { force });
-
-    if (!result.success) {
-      return fail(
-        res,
-        badRequest(result.error ?? "Failed to summarize the job"),
-      );
-    }
-
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      return fail(res, notFound("Job not found"));
-    }
-    ok(res, await hydrateJobPdfFreshness(job));
-
-    if (previousJob) {
-      queueTailoringAutoPdfRegenerationIfNeeded(
-        previousJob,
-        job,
-        "POST /api/jobs/:id/summarize",
-      );
-    }
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * POST /api/jobs/:id/check-sponsor - Check if employer is a visa sponsor
- */
-jobsRouter.post("/:id/check-sponsor", async (req: Request, res: Response) => {
-  try {
-    const job = await jobsRepo.getJobById(req.params.id);
-
-    if (!job) {
-      return fail(res, notFound("Job not found"));
-    }
-
-    if (!job.employer) {
-      return fail(res, badRequest("Job has no employer name"));
-    }
-
-    // Search for sponsor matches
-    const sponsorResults = await visaSponsors.searchSponsors(job.employer, {
-      limit: 10,
-      minScore: 50,
-    });
-
-    const { sponsorMatchScore, sponsorMatchNames } =
-      visaSponsors.calculateSponsorMatchSummary(sponsorResults);
-
-    // Update job with sponsor match info
-    const updatedJob = await jobsRepo.updateJob(job.id, {
-      sponsorMatchScore: sponsorMatchScore,
-      sponsorMatchNames: sponsorMatchNames ?? undefined,
-    });
-
-    if (!updatedJob) {
-      return fail(res, notFound("Job not found"));
-    }
-
-    if (sponsorMatchScore >= 50 && sponsorResults.length > 0) {
-      void trackServerProductEvent(
-        "sponsor_match_found",
-        {
-          match_score: sponsorMatchScore,
-          match_count: sponsorResults.length,
-        },
-        {
-          requestOrigin: resolveRequestOrigin(req),
-          urlPath: "/visa-sponsors",
-        },
-      );
-    }
-
-    ok(res, {
-      ...(await hydrateJobPdfFreshness(updatedJob)),
-      matchResults: sponsorResults.slice(0, 5).map((r) => ({
-        name: r.sponsor.organisationName,
-        score: r.score,
-      })),
-    });
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * POST /api/jobs/:id/generate-pdf - Generate PDF using current manual overrides
- */
-jobsRouter.post("/:id/generate-pdf", async (req: Request, res: Response) => {
-  try {
-    if (isDemoMode()) {
-      const result = await simulateGeneratePdf(req.params.id);
-      if (!result.success) {
-        return fail(
-          res,
-          badRequest(result.error ?? "Failed to generate a resume PDF"),
-        );
-      }
-      const job = await jobsRepo.getJobById(req.params.id);
-      if (!job) {
-        return fail(res, notFound("Job not found"));
-      }
-      return okWithMeta(res, await hydrateJobPdfFreshness(job), {
-        simulated: true,
-      });
-    }
-
-    const result = await generateFinalPdf(req.params.id, {
-      requestOrigin: resolveRequestOrigin(req),
-      analyticsOrigin: "generate_pdf",
-    });
-
-    if (!result.success) {
-      return fail(
-        res,
-        appErrorFromPipelineFailure(result, "Failed to generate a resume PDF"),
-      );
-    }
-
-    const job = await jobsRepo.getJobById(req.params.id);
-    if (!job) {
-      return fail(res, notFound("Job not found"));
-    }
-    ok(res, await hydrateJobPdfFreshness(job));
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * POST /api/jobs/:id/apply - Mark a job as applied
- */
-jobsRouter.post("/:id/apply", async (req: Request, res: Response) => {
-  try {
-    if (isDemoMode()) {
-      const updatedJob = await simulateApplyJob(req.params.id);
-      return okWithMeta(res, await hydrateJobPdfFreshness(updatedJob), {
-        simulated: true,
-      });
-    }
-
-    const job = await jobsRepo.getJobById(req.params.id);
-
-    if (!job) {
-      return fail(res, notFound("Job not found"));
-    }
-
-    const appliedAtDate = new Date();
-    const appliedAt = appliedAtDate.toISOString();
-
-    transitionStage(
-      job.id,
-      "applied",
-      Math.floor(appliedAtDate.getTime() / 1000),
-      {
-        eventLabel: "Applied",
-        actor: "system",
-      },
-      null,
-    );
-
-    const updatedJob = await jobsRepo.updateJob(job.id, {
-      status: "applied",
-      appliedAt,
-    });
-
-    if (updatedJob) {
-      void trackCanonicalActivationEvent(
-        "application_marked_applied",
-        {
-          source: "jobs_apply_route",
-          had_pdf: Boolean(updatedJob.pdfPath),
-          tracer_links_enabled: Boolean(updatedJob.tracerLinksEnabled),
-          sponsor_match_found:
-            typeof updatedJob.sponsorMatchScore === "number" &&
-            updatedJob.sponsorMatchScore >= 50,
-        },
-        {
-          occurredAt: appliedAtDate,
-          requestOrigin: resolveRequestOrigin(req),
-          urlPath: "/jobs",
-        },
-      );
-      notifyJobCompleteWebhook(updatedJob).catch((error) => {
-        logger.warn("Job complete webhook dispatch failed", error);
-      });
-    }
-
-    if (!updatedJob) {
-      return fail(res, notFound("Job not found"));
-    }
-
-    ok(res, await hydrateJobPdfFreshness(updatedJob));
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * DELETE /api/jobs/status/:status - Clear jobs with a specific status
- */
-jobsRouter.delete("/status/:status", async (req: Request, res: Response) => {
-  try {
-    if (isDemoMode()) {
-      return sendDemoBlocked(
-        res,
-        "Clearing jobs by status is disabled to keep the demo stable.",
-        { route: "DELETE /api/jobs/status/:status", status: req.params.status },
-      );
-    }
-
-    const status = req.params.status as JobStatus;
-    const count = await jobsRepo.deleteJobsByStatus(status);
-
-    ok(res, {
-      message: `Cleared ${count} ${status} jobs`,
-      count,
-    });
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
-
-/**
- * DELETE /api/jobs/score/:threshold - Clear jobs with score below threshold (excluding post-apply statuses)
- */
-jobsRouter.delete("/score/:threshold", async (req: Request, res: Response) => {
-  try {
-    if (isDemoMode()) {
-      return sendDemoBlocked(
-        res,
-        "Clearing jobs by score is disabled to keep the demo stable.",
-        {
-          route: "DELETE /api/jobs/score/:threshold",
-          threshold: req.params.threshold,
-        },
-      );
-    }
-
-    const threshold = parseInt(req.params.threshold, 10);
-    if (Number.isNaN(threshold) || threshold < 0 || threshold > 100) {
-      return fail(
-        res,
-        badRequest("Threshold must be a number between 0 and 100"),
-      );
-    }
-
-    const count = await jobsRepo.deleteJobsBelowScore(threshold);
-
-    ok(res, {
-      message: `Cleared ${count} jobs with score below ${threshold}`,
-      count,
-      threshold,
-    });
-  } catch (error) {
-    fail(res, toAppError(error));
-  }
-});
