+++ orchestrator/src/server/infra/request-origin.ts
@@ -1,0 +1,31 @@
+import type { Request } from "express";
+
+export function resolveRequestOrigin(req: Request): string | null {
+  const configuredBaseUrl = process.env.JOBOPS_PUBLIC_BASE_URL?.trim();
+  if (configuredBaseUrl) {
+    try {
+      const parsed = new URL(configuredBaseUrl);
+      if (parsed.protocol && parsed.host) {
+        return `${parsed.protocol}//${parsed.host}`;
+      }
+    } catch {
+      // Ignore invalid env and fall back to request-derived origin.
+    }
+  }
+
+  const trustProxy = Boolean(req.app?.get("trust proxy"));
+  let protocol = (req.protocol || "").trim();
+  let host = (req.header("host") || "").trim();
+
+  if (trustProxy) {
+    const forwardedProto =
+      req.header("x-forwarded-proto")?.split(",")[0]?.trim() ?? "";
+    const forwardedHost =
+      req.header("x-forwarded-host")?.split(",")[0]?.trim() ?? "";
+    if (forwardedProto) protocol = forwardedProto;
+    if (forwardedHost) host = forwardedHost;
+  }
+
+  if (!host || !protocol) return null;
+  return `${protocol}://${host}`;
+}
